### 2025-07-14 #8585 @editorialbot
[PRE REVIEW]: PMMoTo: A Porous Media Morphology and Topology Toolkit
**Submitting author:** @tmweigand (<a href="http://orcid.org/0000-0003-4880-3331">Timothy Weigand</a>)
**Repository:** https://github.com/tmweigand/PMMoTo
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.0
