+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pmmoto"
-version = "0.1.5"
+version = "1.0.0"
 description = "Porous Media Topology and Morphology Toolkit"
 
 license = "MIT"
