### 2025-07-14 755361f
allow mode to be specified for equilibirum distrbutions
