### 2025-07-21 7e0430e
refactoring of merdial axis
