+++ src/pmmoto/filters/medial_axis/medial_axis.py
@@ -1,0 +1,151 @@
+from ._skeletonize_3d_cy import _compute_thin_image
+
+__all__ = ["medial_axis"]
+
+
+def medial_axis(img):
+    ma = img.copy()
+    ma = _compute_thin_image(ma)
+    return ma
+
+
+# class medialAxis(object):
+#     """Calculate Medial Axis and PostProcess
+#     Nodes -> Sets -> Paths
+#     Sets are broken into Reaches -> Medial Nodes -> Medial Clusters
+#     """
+
+#     def __init__(self, Domain, subDomain, grid):
+#         self.Domain = Domain
+#         self.subDomain = subDomain
+#         self.grid = grid
+#         self.Orientation = subDomain.Orientation
+#         self.padding = np.zeros([3], dtype=np.int64)
+#         self.haloGrid = None
+#         self.halo = np.zeros(6)
+#         self.haloPadNeigh = np.zeros(6)
+#         self.haloPadNeighNot = np.zeros(6)
+#         self.MA = None
+
+#     def genMAArrays(self):
+#         """Generate Trimmed MA arrays to get nodeInfo and Correct Neighbor Counts for Boundary Nodes"""
+#         dim = self.MA.shape
+#         tempMA = self.MA[
+#             self.haloPadNeigh[0] : dim[0] - self.haloPadNeigh[1],
+#             self.haloPadNeigh[2] : dim[1] - self.haloPadNeigh[3],
+#             self.haloPadNeigh[4] : dim[2] - self.haloPadNeigh[5],
+#         ]
+
+#         neighMA = np.pad(
+#             self.MA,
+#             (
+#                 (self.haloPadNeighNot[0], self.haloPadNeighNot[1]),
+#                 (self.haloPadNeighNot[2], self.haloPadNeighNot[3]),
+#                 (self.haloPadNeighNot[4], self.haloPadNeighNot[5]),
+#             ),
+#             "constant",
+#             constant_values=0,
+#         )
+#         return tempMA, neighMA
+
+
+# def medialAxisEval(subDomain, porousMedia, grid, distance, connect=False, trim=False):
+
+#     rank = subDomain.ID
+#     size = subDomain.Domain.numSubDomains
+
+#     grid = porousMedia.grid
+
+#     ### Initialize Classes
+#     sDMA = medialAxis(Domain=subDomain.Domain, subDomain=subDomain, grid=grid)
+
+#     ### Extract MA
+#     mE = medialExtraction.medialExtraction(
+#         Domain=subDomain.Domain, subDomain=subDomain, grid=grid, edt=distance
+#     )
+
+#     if not connect:
+#         sDMA.MA = mE.extractMedialAxis(connect)
+
+#     if connect:
+
+#         ### CLEAN this up with Array and MA Neighbors
+#         sDMA.MA, sDMA.haloPadNeigh, sDMA.haloPadNeighNot = mE.extractMedialAxis(connect)
+
+#         ### Get Info for Medial Axis Nodes and Get Connected Sets and Boundary Sets
+#         tempMA, neighMA = sDMA.genMAArrays()
+
+#         Nodes = nodes.get_node_info(
+#             rank,
+#             tempMA,
+#             1,
+#             porousMedia.inlet,
+#             porousMedia.outlet,
+#             subDomain.Domain,
+#             porousMedia.loopInfo,
+#             subDomain,
+#             subDomain.Orientation,
+#         )
+#         maNodesType = medialNodes.updateMANeighborCount(
+#             neighMA, porousMedia, subDomain.Orientation, Nodes[0]
+#         )
+#         sDMA.MA = np.ascontiguousarray(tempMA)
+
+#         mSets = medialNodes.getConnectedMedialAxis(
+#             subDomain, sDMA.MA, Nodes, maNodesType
[…]
