+++ src/pmmoto/filters/medial_axis/__init__.py
@@ -1,0 +1,6 @@
+"""Medial axis filter for PMMoTo."""
+
+from .medial_axis import medial_axis
+from . import _medial_axis
+
+__all__ = ["medial_axis", "_medial_axis"]

[…]
