+++ _sources/examples/index.rst.txt
@@ -1,0 +1,12 @@
+Examples
+========
+
+Examples demonstrating how to use PMMoTo for porous media analysis:
+
+.. toctree::
+   :maxdepth: 1
+
+   drainage_inkbottle
+   sphere_pack_psd
+   connected_pathways
+   md_porous_media
