+++ src/pmmoto/core/features.py
@@ -8,6 +8,7 @@ from typing import TYPE_CHECKING
 import numpy as np
 from numpy.typing import NDArray
 from .boundary_types import BoundaryType
+from .octants import get_neighbors_vertices
 from .orientation import get_boundary_id
 from .orientation import FEATURE_MAP
 
@@ -27,7 +28,7 @@ class Feature(object):
         dim: int,
         feature_id: tuple[int, ...],
         neighbor_rank: int,
-        boundary_type: BoundaryType,
+        boundary_type: BoundaryType | tuple[BoundaryType, ...],
         global_boundary: bool | None = None,
     ):
         """Initialize a Feature.
@@ -49,8 +50,8 @@ class Feature(object):
         self.periodic = False
         self.periodic_correction: tuple[int, ...] = (0, 0, 0)
         self.extend: tuple[tuple[int, int], ...] = ((0, 0), (0, 0), (0, 0))
-        self.own: NDArray[np.uint64] = np.zeros((self.dim, 2), dtype=np.uint64)
-        self.neighbor: NDArray[np.uint64] = np.zeros((self.dim, 2), dtype=np.uint64)
+        self.own: None | NDArray[np.uint64] = None
+        self.neighbor: None | NDArray[np.uint64] = None
 
     def convert_feature_id(self, index: tuple[int, ...] | None = None) -> int:
         """Convert the feature type to id.
@@ -64,39 +65,66 @@ class Feature(object):
 
         return get_boundary_id(index)
 
-    def is_periodic(self, boundary_type: BoundaryType) -> bool:
+    def is_periodic(
+        self, boundary_type: BoundaryType | tuple[BoundaryType, ...]
+    ) -> bool:
         """Determine if a feature is periodic
 
         Returns:
             bool: True if periodic
 
         """
+        if isinstance(boundary_type, tuple):
+            return all(bt == BoundaryType.PERIODIC for bt in boundary_type)
+
         return boundary_type == BoundaryType.PERIODIC
+
+    def set_voxels(
+        self,
+        voxels: tuple[int, ...],
+        pad: tuple[tuple[int, int], ...] = ((0, 0), (0, 0), (0, 0)),
+        extract_features: bool = False,
+    ):
+        """Set attributes for own and neighbor voxels"""
+        self.own, self.neighbor = self.get_voxels(voxels, pad, extract_features)
 
     def get_voxels(
         self,
         voxels: tuple[int, ...],
         pad: tuple[tuple[int, int], ...] = ((0, 0), (0, 0), (0, 0)),
-    ) -> None:
-        """Determine the voxel indices for the feature"""
+        extract_features: bool = False,
+    ) -> tuple[NDArray[np.uint64], NDArray[np.uint64]]:
+        """Determine the voxel indices for the feature.
+
+        If pad is zeros and extract_features is True, the features are decomposed
+        """
+        own: NDArray[np.uint64] = np.zeros((self.dim, 2), dtype=np.uint64)
+        neighbor: NDArray[np.uint64] = np.zeros((self.dim, 2), dtype=np.uint64)
+
         padded = np.any(pad)
-        for i in range(self.dim):
-            f_id = self.feature_id[i]
+        for i, f_id in enumerate(self.feature_id):
             length = voxels[i]
             p_before, p_after = pad[i]
             if f_id == -1:
-                self.own[i] = self.compute_lower_face(p_before)
+                own[i] = self.compute_lower_face(p_before)
                 if padded:
-                    self.neighbor[i] = self.compute_lower_neighbor(p_before)
+                    neighbor[i] = self.compute_lower_neighbor(p_before)
             elif f_id == 1:
-                self.own[i] = self.compute_upper_face(length, p_after)
[…]
