+++ tests/core/test_features.py
@@ -1,5 +1,7 @@
 """test_features.py"""
 
+import numpy as np
+import pytest
 import pmmoto
 
 
@@ -9,12 +11,17 @@ def test_feature():
     neighbor_rank = 0
     boundary_type = pmmoto.BoundaryType.END
 
-    _ = pmmoto.core.features.Feature(
+    # Non_feature because all are 0
+    non_feature = pmmoto.core.features.Feature(
         dim=3,
         feature_id=feature_id,
         neighbor_rank=neighbor_rank,
         boundary_type=boundary_type,
     )
+
+    assert non_feature.dim == 3
+    assert non_feature.neighbor_rank == neighbor_rank
+    assert non_feature.boundary_type == boundary_type
 
 
 def test_face():
@@ -28,71 +35,356 @@ def test_face():
     )
 
     assert feature.info.opp == (-1, 0, 0)
-    assert feature.boundary_type == "periodic"
+    assert feature.boundary_type == pmmoto.BoundaryType.PERIODIC
     assert feature.get_periodic_correction() == (-1, 0, 0)
 
     feature_id = (0, -1, 0)
-    neighbor_rank = [0]
-    boundary_type = "end"
+    neighbor_rank = 0
+    boundary_type = pmmoto.BoundaryType.END
 
     feature = pmmoto.core.subdomain_features.Face(
         feature_id=feature_id, neighbor_rank=neighbor_rank, boundary_type=boundary_type
     )
 
     assert feature.info.opp == (0, 1, 0)
-    assert feature.boundary_type == "end"
+    assert feature.boundary_type == pmmoto.BoundaryType.END
     assert feature.get_periodic_correction() == (0, 0, 0)
 
 
 def test_edge():
     """Test the edge feature class"""
     feature_id = (1, 1, 0)
-    neighbor_rank = [0]
-    boundary_type = "periodic"
+    neighbor_rank = 0
+    boundary_type = [pmmoto.BoundaryType.INTERNAL, pmmoto.BoundaryType.PERIODIC]
 
     feature = pmmoto.core.subdomain_features.Edge(
         feature_id=feature_id, neighbor_rank=neighbor_rank, boundary_type=boundary_type
     )
 
     assert feature.info.opp == (-1, -1, 0)
-    assert feature.boundary_type == "periodic"
-    assert feature.get_periodic_correction() == (-1, -1, 0)
+    assert feature.boundary_type == (
+        pmmoto.BoundaryType.INTERNAL,
+        pmmoto.BoundaryType.PERIODIC,
+    )
+    assert feature.get_periodic_correction() == (0, 0, 0)
 
     feature_id = (-1, 0, -1)
-    neighbor_rank = [0, 2]
-    boundary_type = "end"
+    neighbor_rank = 0
+    boundary_type = [pmmoto.BoundaryType.END, pmmoto.BoundaryType.END]
 
     feature = pmmoto.core.subdomain_features.Edge(
         feature_id=feature_id, neighbor_rank=neighbor_rank, boundary_type=boundary_type
     )
 
     assert feature.info.opp == (1, 0, 1)
-    assert feature.boundary_type == "end"
+    assert feature.boundary_type == (pmmoto.BoundaryType.END, pmmoto.BoundaryType.END)
     assert feature.get_periodic_correction() == (0, 0, 0)
 
 
 def test_corner():
     """Test the corner feature class"""
     feature_id = (1, 1, -1)
-    neighbor_rank = [0]
-    boundary_type = "periodic"
+    neighbor_rank = 0
+    boundary_type = [
+        pmmoto.BoundaryType.PERIODIC,
[…]
