### 2025-09-09 42592db
fixes volume calculation for bins
