### 2025-09-09 be6e865
specify osx 14 for testing. not ready for 15.
