+++ .github/workflows/tests.yml
@@ -10,7 +10,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macOS-latest]
+        os: [ubuntu-latest, macos-14]
         python-version: ["3.10", "3.11", "3.12"]
         mpi: [mpich, openmpi]
 
