### 2025-09-09 941b7ac
modify lammps read to include id
