### 2025-10-23 f8bf7ab
speeds up reading of LAMMPS files
