+++ src/pmmoto/io/data_read.hpp
@@ -3,6 +3,7 @@
 
 #include <zlib.h>
 
+#include <cstdint>
 #include <fstream>
 #include <map>
 #include <memory>
@@ -12,13 +13,22 @@
 #include <string_view>
 #include <vector>
 
+// struct LammpsData
+// {
+//     std::vector<std::vector<double> > atom_positions;
+//     std::vector<uint8_t> atom_types;
+//     std::vector<uint64_t> atom_ids;
+//     std::vector<std::vector<double> > domain_data;
+//     double timestep;
+// };
+
 struct LammpsData
 {
-    std::vector<std::vector<double> > atom_positions;
-    std::vector<int> atom_types;
+    std::vector<double> atom_positions;
+    std::vector<uint64_t> atom_ids;
+    std::vector<uint8_t> atom_types;
     std::vector<std::vector<double> > domain_data;
-    std::vector<int> atom_ids;
-    double timestep;
+    double timestep = 0.0;
 };
 
 class LammpsReader
@@ -27,7 +37,6 @@ private:
     using AtomIdMap =
         std::map<std::pair<int, double>, int>;       // type, charge -> atom_id
     static constexpr size_t BUFFER_SIZE = 64 * 1024; // 64KB buffer
-    static inline std::vector<std::string> tokens;
 
     class GzFileWrapper
     {
@@ -61,27 +70,6 @@ private:
         GzFileWrapper(const GzFileWrapper&) = delete;
         GzFileWrapper& operator=(const GzFileWrapper&) = delete;
     };
-
-    // Efficient string splitting
-    static void split(const std::string& str, std::vector<std::string>& out)
-    {
-        out.clear();
-        size_t start = 0;
-        size_t end = 0;
-
-        // Pre-reserve some space to avoid reallocations
-        out.reserve(8); // Most lines have 8 or fewer tokens
-
-        while ((end = str.find(' ', start)) != std::string::npos)
-        {
-            if (end > start) out.emplace_back(str.substr(start, end - start));
-            start = end + 1;
-            // Skip multiple spaces
-            while (start < str.length() && str[start] == ' ') ++start;
-        }
-
-        if (start < str.length()) out.emplace_back(str.substr(start));
-    }
 
     static std::string readGzipFile(const std::string& filename)
     {
@@ -146,15 +134,6 @@ private:
         return std::stod(line);
     }
 
-    static int readNumAtoms(const std::string& line, LammpsData& data)
-    {
-        int num_atoms = std::stoi(line);
-        data.atom_positions.resize(num_atoms, std::vector<double>(3, 0.0));
-        data.atom_ids.resize(num_atoms, 0);
-        data.atom_types.resize(num_atoms, 0);
-        return num_atoms;
-    }
-
     static void
     readDomainBounds(const std::string& line, int dim, LammpsData& data)
     {
@@ -166,137 +145,148 @@ private:
         }
     }
 
-    static void processAtomLine(const std::string& line,
-                                size_t atom_count,
-                                LammpsData& data)
+    static inline void processAtomLine(const std::string_view& line,
+                                       size_t atom_count,
+                                       LammpsData& data)
     {
-        split(line, tokens);
+        size_t pos = 0;
+        size_t next;
 
-        if (tokens.size() >= 8)
+        // Parse id
+        next = line.find(' ', pos);
+        data.atom_ids[atom_count] =
+            std::stoull(std::string(line.substr(pos, next - pos)));
+        pos = next + 1;
+
+        // Parse mol (skip)
+        next = line.find(' ', pos);
+        pos = next + 1;
+
+        // Parse type
+        next = line.find(' ', pos);
+        data.atom_types[atom_count] =
[…]
