### 2025-10-23 2787669
speeds up reading of LAMMPS files
