+++ src/pmmoto/io/_data_read.pyx
@@ -1,48 +1,83 @@
 import numpy as np
 cimport numpy as np
 from libcpp.string cimport string
-from libcpp.vector cimport vector
 from libcpp.map cimport map
 from libcpp.pair cimport pair
+from libc.stdint cimport uint64_t,uint8_t
 
-# Initialize numpy
 np.import_array()
 
+cdef class AtomIdMapWrapper:
+    def __cinit__(self):
+        self._c_map_ptr = new AtomIdMap()
+
+    def __dealloc__(self):
+        if self._c_map_ptr is not NULL:
+            del self._c_map_ptr
+            self._c_map_ptr = NULL
+
+# Convert Python dict[(int,float)->int] -> C++ AtomIdMap (returned by value)
+cdef AtomIdMap dict_to_atom_id_map(object py_map):
+    cdef AtomIdMap c_map
+    cdef object py_items = py_map.items()
+    cdef object py_key, py_val
+    cdef int k_type
+    cdef double k_charge
+    cdef int v
+    cdef pair[int, double] pkey
+    cdef pair[pair[int, double], int] entry
+    for py_key, py_val in py_items:
+        # expect py_key to be a (int, float) tuple
+        k_type = <int>py_key[0]
+        k_charge = <double>py_key[1]
+        v = <int>py_val
+        pkey = pair[int, double](k_type, k_charge)
+        entry = pair[pair[int, double], int](pkey, v)
+        c_map.insert(entry)
+    return c_map
+
 def read_lammps_atoms(str filename, type_map=None):
-    """
-    Read LAMMPS atom data from file
-    
-    Args:
-        filename (str): Path to LAMMPS data file
-        type_map (dict, optional): Mapping of (type, charge) pairs to new types
-            Example: {(1, 0.4): 2, (1, -0.4): 3}
-        
-    Returns:
-        tuple: (positions, types, domain_data, timestep)
-            - positions: np.ndarray of shape (n_atoms, 3)
-            - types: np.ndarray of shape (n_atoms,)
-            - domain_data: np.ndarray of shape (3, 2)
-            - timestep: float
-    """
-    # Convert Python string to C++ string
     cdef string cpp_filename = filename.encode('utf-8')
-    
-    # Convert Python dict to C++ map if provided
-    cdef map[pair[int, double], int] cpp_type_map
-    if type_map is not None:
-        cpp_type_map = type_map
-    
-    # Call C++ function
     cdef LammpsData data
+    cdef const AtomIdMap* _type_map = NULL
+    cdef AtomIdMap tmp_map
+    cdef AtomIdMapWrapper wrapper
+
     if type_map is not None:
-        data = LammpsReader.read_lammps_atoms_with_map(cpp_filename, cpp_type_map)
-    else:
-        data = LammpsReader.read_lammps_atoms(cpp_filename)
-    
-    # Convert C++ vectors to numpy arrays
-    cdef np.ndarray[double, ndim=2] positions = np.array(data.atom_positions)
-    cdef np.ndarray[long, ndim=1] types = np.array(data.atom_types)
-    cdef np.ndarray[long, ndim=1] ids = np.array(data.atom_ids)
-    cdef np.ndarray[double, ndim=2] domain = np.array(data.domain_data)
-    
+        # if user passed our wrapper, use its pointer directly
+        if isinstance(type_map, AtomIdMapWrapper):
+            wrapper = <AtomIdMapWrapper> type_map
+            _type_map = wrapper._c_map_ptr
+        else:
+            # convert Python dict to C++ map and pass pointer to it
[…]
