### 2025-10-25 e9033d1
add build optimizations
