+++ setup.py
@@ -11,21 +11,29 @@ import numpy
 
 Cython.Compiler.Options.annotate = True
 
-extra_compile_args = []
-
 if sys.platform == "win32":
-    # MSVC
-    extra_compile_args += ["/std:c++17", "/O2"]
-else:
-    # GCC / Clang
-    extra_compile_args += ["-std=c++17", "-O3", "-pthread"]
+    sys.exit("Windows is not supported for building this package.")
+
+# common fast flags for Unix / macOS (no language-specific flags)
+base_compile_args = [
+    "-O3",
+    "-ffast-math",
+    "-funroll-loops",
+    "-fomit-frame-pointer",
+    "-fno-math-errno",
+    "-flto",
+    "-pthread",
+]
+
+# C++-specific flags (only add to extensions using C++)
+cpp_compile_args = ["-std=c++17"] + base_compile_args[:]
+
+extra_link_args = ["-flto"]
 
 if sys.platform == "darwin":
-    # Use modern libc++; do NOT pin to ancient macOS
-    extra_compile_args += ["-stdlib=libc++"]
-
-
-cmdclass = {}
+    base_compile_args += ["-stdlib=libc++", "-mmacosx-version-min=10.9"]
+    cpp_compile_args += ["-stdlib=libc++", "-mmacosx-version-min=10.9"]
+    extra_link_args += ["-stdlib=libc++", "-mmacosx-version-min=10.9"]
 
 ext_modules = [
     Extension(
@@ -33,14 +41,16 @@ ext_modules = [
         ["src/pmmoto/core/_voxels.pyx"],
         include_dirs=["src/pmmoto/core"],
         language="c++",
-        extra_compile_args=extra_compile_args,
+        extra_compile_args=cpp_compile_args,
+        extra_link_args=extra_link_args,
     ),
     Extension(
         "pmmoto.filters.distance._distance",
         ["src/pmmoto/filters/distance/_distance.pyx"],
         include_dirs=["src/pmmoto/filters/distance", numpy.get_include()],
         language="c++",
-        extra_compile_args=extra_compile_args,
+        extra_compile_args=cpp_compile_args,
+        extra_link_args=extra_link_args,
     ),
     Extension(
         "pmmoto.analysis._minkowski",
@@ -50,6 +60,9 @@ ext_modules = [
             "src/pmmoto/analysis/minkowskic.c",
         ],
         include_dirs=["src/pmmoto/analysis"],
+        language="c",  # keep C for the .c sources
+        extra_compile_args=base_compile_args,
+        extra_link_args=extra_link_args,
     ),
     Extension(
         "pmmoto.io._data_read",
@@ -57,37 +70,43 @@ ext_modules = [
         include_dirs=["src/pmmoto/io"],
         libraries=["z"],
         language="c++",
-        extra_compile_args=extra_compile_args,
+        extra_compile_args=cpp_compile_args,
+        extra_link_args=extra_link_args,
     ),
     Extension(
         "pmmoto.particles._particles",
         ["src/pmmoto/particles/_particles.pyx"],
         include_dirs=["src/pmmoto/particles/"],
         language="c++",
-        extra_compile_args=extra_compile_args,
+        extra_compile_args=cpp_compile_args,
+        extra_link_args=extra_link_args,
     ),
     Extension(
         "pmmoto.domain_generation._rdf",
         ["src/pmmoto/domain_generation/_rdf.pyx"],
         include_dirs=["src/pmmoto/domain_generation/", "src/pmmoto/particles/"],
         language="c++",
-        extra_compile_args=extra_compile_args,
+        extra_compile_args=cpp_compile_args,
+        extra_link_args=extra_link_args,
     ),
     Extension(
[…]
