### 2025-10-25 0ae6194
add build optimizations
