### 2025-10-25 e71080c
adds fast_float dependency and adjusts atom_read
