+++ src/pmmoto/io/data_read.hpp
@@ -1,9 +1,13 @@
 #ifndef DATA_READ_HPP
 #define DATA_READ_HPP
 
+#include "fast_float/fast_float.h" // add fast_float to project (git submodule / fetch)
+
 #include <zlib.h>
 
+#include <charconv>
 #include <cstdint>
+#include <cstdlib>
 #include <fstream>
 #include <map>
 #include <memory>
@@ -12,15 +16,6 @@
 #include <string>
 #include <string_view>
 #include <vector>
-
-// struct LammpsData
-// {
-//     std::vector<std::vector<double> > atom_positions;
-//     std::vector<uint8_t> atom_types;
-//     std::vector<uint64_t> atom_ids;
-//     std::vector<std::vector<double> > domain_data;
-//     double timestep;
-// };
 
 struct LammpsData
 {
@@ -145,6 +140,79 @@ private:
         }
     }
 
+    static inline const char* skip_spaces(const char* p, const char* end)
+    {
+        while (p < end && (*p == ' ' || *p == '\t')) ++p;
+        return p;
+    }
+
+    static inline const char*
+    parse_uint64_ptr(const char* p, const char* end, uint64_t& out)
+    {
+        auto res = std::from_chars(p, end, out);
+        return res.ptr;
+    }
+
+    static inline const char*
+    parse_int_ptr(const char* p, const char* end, int& out)
+    {
+        auto res = std::from_chars(p, end, out);
+        return res.ptr;
+    }
+
+    static inline const char*
+    parse_double_ptr(const char* p, const char* end, double& out)
+    {
+        // fast_float::from_chars is much faster than std::stod/strtod
+        auto result = fast_float::from_chars(p, end, out);
+        return result.ptr;
+    }
+
+    // Replace processAtomLine with a pointer-based routine used in the main
+    // loop
+    static inline const char* parse_atom_fields(const char* p,
+                                                const char* end,
+                                                uint64_t& id_out,
+                                                int& type_out,
+                                                double coords[3],
+                                                double& charge_out,
+                                                bool parse_charge)
+    {
+        p = skip_spaces(p, end);
+        p = parse_uint64_ptr(p, end, id_out);
+        p = skip_spaces(p, end);
+
+        // mol (skip token)
+        while (p < end && *p != ' ' && *p != '\t' && *p != '\n') ++p;
+        p = skip_spaces(p, end);
+
+        p = parse_int_ptr(p, end, type_out);
+        p = skip_spaces(p, end);
+
+        // mass (skip token)
+        while (p < end && *p != ' ' && *p != '\t' && *p != '\n') ++p;
+        p = skip_spaces(p, end);
+
+        if (parse_charge)
+        {
+            p = parse_double_ptr(p, end, charge_out);
+            p = skip_spaces(p, end);
+        }
+        else
+        {
+            // skip q token
+            while (p < end && *p != ' ' && *p != '\t' && *p != '\n') ++p;
+            p = skip_spaces(p, end);
+        }
+
+        for (int i = 0; i < 3; ++i)
+        {
+            p = parse_double_ptr(p, end, coords[i]);
+            p = skip_spaces(p, end);
+        }
+        return p;
+    }
+
     static inline void processAtomLine(const std::string_view& line,
                                        size_t atom_count,
                                        LammpsData& data)
@@ -245,48 +313,120 @@ public:
         LammpsData data;
         data.domain_data = { { { 0.0, 0.0 }, { 0.0, 0.0 }, { 0.0, 0.0 } } };
 
-        auto file = openFile(filename);
-        std::string line;
+        // Read whole file (gz or plain) into a single string buffer
+        std::string content;
+        if (filename.size() > 3 &&
+            filename.compare(filename.size() - 3, 3, ".gz") == 0)
+        {
[…]
