### 2025-10-25 9cac6c2
fixes fast_float dependency
