### 2025-10-25 6c4c076
add to manifest
