### 2025-10-25 8312a83
edit manifest
