### 2025-10-25 f7bbe11
more read optimizations
