+++ src/pmmoto/io/data_read.hpp
@@ -1,436 +1,423 @@
 #ifndef DATA_READ_HPP
 #define DATA_READ_HPP
 
+// You vendored fast_float headers under extern/fast_float/include
+// Ensure setup.py adds that path to include_dirs.
 #include "fast_float.h"
+#include <sys/mman.h>
+#include <sys/stat.h>
 
+#include <fcntl.h>
+#include <unistd.h>
 #include <zlib.h>
 
 #include <charconv>
 #include <cstdint>
 #include <cstdlib>
+#include <cstring>
 #include <fstream>
 #include <map>
-#include <memory>
-#include <sstream>
 #include <stdexcept>
 #include <string>
 #include <string_view>
+#include <unordered_map>
+#include <utility>
 #include <vector>
 
+// Result container (must match Cython .pxd)
 struct LammpsData
 {
-    std::vector<double> atom_positions;
     std::vector<uint64_t> atom_ids;
     std::vector<uint8_t> atom_types;
-    std::vector<std::vector<double> > domain_data;
+    std::vector<double> atom_positions;            // 3N flat
+    std::vector<std::vector<double> > domain_data; // [[lo,hi],[lo,hi],[lo,hi]]
     double timestep = 0.0;
 };
 
-class LammpsReader
+// ====== Fast IO helpers ======
+
+// RAII mmap view (no copy for plain files)
+struct MMapBuffer
 {
-private:
-    using AtomIdMap =
-        std::map<std::pair<int, double>, int>;       // type, charge -> atom_id
-    static constexpr size_t BUFFER_SIZE = 64 * 1024; // 64KB buffer
+    const char* ptr = nullptr;
+    size_t len = 0;
+    void* map = nullptr;
 
-    class GzFileWrapper
+    MMapBuffer() = default;
+    MMapBuffer(const MMapBuffer&) = delete;
+    MMapBuffer& operator=(const MMapBuffer&) = delete;
+
+    // Move constructor and assignment
+    MMapBuffer(MMapBuffer&& other) noexcept : ptr(other.ptr),
+                                              len(other.len),
+                                              map(other.map)
     {
-        gzFile file_;
-
-    public:
-        explicit GzFileWrapper(const std::string& filename)
-        {
-            file_ = gzopen(filename.c_str(), "r");
-            if (!file_)
-            {
-                throw std::runtime_error("Could not open gzipped file: " +
-                                         filename);
-            }
-        }
-
-        ~GzFileWrapper()
-        {
-            if (file_)
-            {
-                gzclose(file_);
-            }
-        }
-
-        gzFile get()
-        {
-            return file_;
-        }
-
-        // Prevent copying
-        GzFileWrapper(const GzFileWrapper&) = delete;
-        GzFileWrapper& operator=(const GzFileWrapper&) = delete;
-    };
-
-    static std::string readGzipFile(const std::string& filename)
+        other.ptr = nullptr;
+        other.len = 0;
+        other.map = nullptr;
+    }
+    MMapBuffer& operator=(MMapBuffer&& other) noexcept
     {
-        GzFileWrapper file(filename); // RAII wrapper handles cleanup
-
-        std::string content;
-        content.reserve(1024 * 1024);
-        std::vector<char> buffer(BUFFER_SIZE);
-
-        while (int bytesRead = gzread(file.get(), buffer.data(), buffer.size()))
+        if (this != &other)
         {
-            if (bytesRead < 0)
-            {
-                throw std::runtime_error("Error reading gzipped file: " +
-                                         filename);
-            }
-            content.append(buffer.data(), bytesRead);
+            if (map && len) munmap(map, len);
+            ptr = other.ptr;
+            len = other.len;
+            map = other.map;
+            other.ptr = nullptr;
+            other.len = 0;
+            other.map = nullptr;
         }
-
[…]
