### 2025-10-25 6fa19b9
more yet read optimizations
