+++ src/pmmoto/io/data_read.hpp
@@ -24,6 +24,15 @@
 #include <utility>
 #include <vector>
 
+// Branch hints for hot paths
+#if defined(__GNUC__) || defined(__clang__)
+#define LIKELY(x) __builtin_expect(!!(x), 1)
+#define UNLIKELY(x) __builtin_expect(!!(x), 0)
+#else
+#define LIKELY(x) (x)
+#define UNLIKELY(x) (x)
+#endif
+
 // Result container (must match Cython .pxd)
 struct LammpsData
 {
@@ -85,32 +94,35 @@ map_file_readonly(const std::string& filename)
 {
     MMapBuffer out;
     int fd = open(filename.c_str(), O_RDONLY);
-    if (fd < 0) throw std::runtime_error("open failed: " + filename);
+    if (UNLIKELY(fd < 0)) throw std::runtime_error("open failed: " + filename);
 
     struct stat st
     {
     };
-    if (fstat(fd, &st) != 0)
+    if (UNLIKELY(fstat(fd, &st) != 0))
     {
         int e = errno;
         close(fd);
         throw std::runtime_error("fstat failed (" + std::to_string(e) +
                                  "): " + filename);
     }
+
     size_t size = static_cast<size_t>(st.st_size);
     if (size == 0)
     {
         close(fd);
         return out;
     }
+
     void* p = mmap(nullptr, size, PROT_READ, MAP_PRIVATE, fd, 0);
     int e = errno;
     close(fd);
-    if (p == MAP_FAILED)
+    if (UNLIKELY(p == MAP_FAILED))
     {
         throw std::runtime_error("mmap failed (" + std::to_string(e) +
                                  "): " + filename);
     }
+
     out.map = p;
     out.ptr = static_cast<const char*>(p);
     out.len = size;
@@ -122,14 +134,19 @@ static inline size_t
 gzip_uncompressed_size_from_footer(const std::string& filename)
 {
     std::ifstream ifs(filename, std::ios::binary | std::ios::ate);
-    if (!ifs) throw std::runtime_error("Could not open gz file: " + filename);
+    if (UNLIKELY(!ifs))
+        throw std::runtime_error("Could not open gz file: " + filename);
+
     auto end = ifs.tellg();
-    if (end < std::streamoff(4))
+    if (UNLIKELY(end < std::streamoff(4)))
         throw std::runtime_error("Invalid gz (too small): " + filename);
+
     ifs.seekg(end - std::streamoff(4));
     unsigned char isize[4]{};
     ifs.read(reinterpret_cast<char*>(isize), 4);
-    if (!ifs) throw std::runtime_error("Failed reading gz footer: " + filename);
+    if (UNLIKELY(!ifs))
+        throw std::runtime_error("Failed reading gz footer: " + filename);
+
     size_t n = size_t(isize[0]) | (size_t(isize[1]) << 8) |
                (size_t(isize[2]) << 16) | (size_t(isize[3]) << 24);
     return n;
@@ -140,13 +157,15 @@ static inline std::string
 readGzipFile_fast_zlib(const std::string& filename)
 {
     std::ifstream ifs(filename, std::ios::binary | std::ios::ate);
-    if (!ifs) throw std::runtime_error("open gz failed: " + filename);
+    if (UNLIKELY(!ifs)) throw std::runtime_error("open gz failed: " + filename);
+
     size_t comp_size = static_cast<size_t>(ifs.tellg());
     ifs.seekg(0);
+
     std::string comp;
     comp.resize(comp_size);
     ifs.read(comp.data(), comp_size);
-    if (!ifs) throw std::runtime_error("read gz failed: " + filename);
+    if (UNLIKELY(!ifs)) throw std::runtime_error("read gz failed: " + filename);
[…]
