+++ src/pmmoto/io/data_read.hpp
@@ -11,6 +11,7 @@
 #include <unistd.h>
 #include <zlib.h>
 
+#include <algorithm>
 #include <charconv>
 #include <cstdint>
 #include <cstdlib>
@@ -20,32 +21,41 @@
 #include <stdexcept>
 #include <string>
 #include <string_view>
+#include <thread>
 #include <unordered_map>
 #include <utility>
 #include <vector>
 
-// Branch hints for hot paths
+// SIMD headers for fast scanning
+#ifdef __SSE2__
+#include <emmintrin.h>
+#endif
+#ifdef __AVX2__
+#include <immintrin.h>
+#endif
+
+// Branch hints
 #if defined(__GNUC__) || defined(__clang__)
 #define LIKELY(x) __builtin_expect(!!(x), 1)
 #define UNLIKELY(x) __builtin_expect(!!(x), 0)
+#define PREFETCH_READ(addr) __builtin_prefetch(addr, 0, 3)
+#define PREFETCH_WRITE(addr) __builtin_prefetch(addr, 1, 3)
 #else
 #define LIKELY(x) (x)
 #define UNLIKELY(x) (x)
+#define PREFETCH_READ(addr) ((void)0)
+#define PREFETCH_WRITE(addr) ((void)0)
 #endif
 
-// Result container (must match Cython .pxd)
 struct LammpsData
 {
     std::vector<uint64_t> atom_ids;
     std::vector<uint8_t> atom_types;
-    std::vector<double> atom_positions;            // 3N flat
-    std::vector<std::vector<double> > domain_data; // [[lo,hi],[lo,hi],[lo,hi]]
+    std::vector<double> atom_positions;
+    std::vector<std::vector<double> > domain_data;
     double timestep = 0.0;
 };
 
-// ====== Fast IO helpers ======
-
-// RAII mmap view (no copy for plain files)
 struct MMapBuffer
 {
     const char* ptr = nullptr;
@@ -56,7 +66,6 @@ struct MMapBuffer
     MMapBuffer(const MMapBuffer&) = delete;
     MMapBuffer& operator=(const MMapBuffer&) = delete;
 
-    // Move constructor and assignment
     MMapBuffer(MMapBuffer&& other) noexcept : ptr(other.ptr),
                                               len(other.len),
                                               map(other.map)
@@ -65,6 +74,7 @@ struct MMapBuffer
         other.len = 0;
         other.map = nullptr;
     }
+
     MMapBuffer& operator=(MMapBuffer&& other) noexcept
     {
         if (this != &other)
@@ -82,10 +92,7 @@ struct MMapBuffer
 
     ~MMapBuffer()
     {
-        if (map && len)
-        {
-            munmap(map, len);
-        }
+        if (map && len) munmap(map, len);
     }
 };
 
@@ -123,13 +130,15 @@ map_file_readonly(const std::string& filename)
                                  "): " + filename);
     }
 
+    // Advise kernel we'll read sequentially
+    madvise(p, size, MADV_SEQUENTIAL);
+
     out.map = p;
     out.ptr = static_cast<const char*>(p);
     out.len = size;
     return out;
 }
 
-// Read last 4 bytes of gzip (ISIZE) to get uncompressed size (mod 2^32)
 static inline size_t
 gzip_uncompressed_size_from_footer(const std::string& filename)
 {
@@ -152,7 +161,6 @@ gzip_uncompressed_size_from_footer(const std::string& filename)
     return n;
 }
 
-// Single-pass zlib inflate using exact-size allocation from ISIZE
 static inline std::string
 readGzipFile_fast_zlib(const std::string& filename)
 {
[…]
