### 2025-10-25 31e8d6f
platform specs
