### 2025-10-25 61f1487
intel mac issues
