+++ .github/workflows/lint.yml
@@ -1,31 +1,35 @@
 name: Lint
 
 on:
-  - push
-  - pull_request
+  push:
+    branches: [main]
+    tags: ["v*"]
+  pull_request:
+    branches: [main]
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.head.sha || github.sha }}
+  cancel-in-progress: true
+
+permissions:
+  contents: read
 
 jobs:
   lint:
     runs-on: ${{ matrix.os }}
+    timeout-minutes: 30
     strategy:
       matrix:
         os: [macos-14]
         python-version: ["3.11"]
 
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install MPI
-        run: |
-          if [ "$RUNNER_OS" = "Linux" ]; then
-            sudo apt update
-            sudo apt install -y libmpich-dev zlib1g-dev
-          elif [ "$RUNNER_OS" = "macOS" ]; then
-            brew install mpich zlib
-          fi
+          cache: "pip"
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
