### 2025-10-25 a042d5e
clean up CI
