+++ MANIFEST.in
@@ -2,4 +2,4 @@ include README.md
 include LICENSE
 recursive-include examples *
 recursive-include src/ *.pyx *.pxd *.py *.hpp *.h
-recursive-include extern/ *.pyx *.pxd *.py *.hpp *.h
+recursive-include extern/ *.h
