### 2025-10-26 d13f825
add back mpi
