+++ .github/workflows/Coverage.yaml
@@ -31,9 +31,15 @@ jobs:
           python-version: ${{ matrix.python-version }}
           cache: "pip"
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install tox tox-gh-actions
+              run: |
+                if [ "$RUNNER_OS" = "Linux" ]; then
+                  sudo apt update
+                  sudo apt install -y libmpich-dev zlib1g-dev
+                elif [ "$RUNNER_OS" = "macOS" ]; then
+                  brew install mpich zlib
+                fi
+                python -m pip install --upgrade pip
+                pip install tox tox-gh-actions
       - name: Run coverage
         run: tox -e coverage
       - name: Upload coverage report
