+++ .github/workflows/lint.yml
@@ -31,8 +31,14 @@ jobs:
           python-version: ${{ matrix.python-version }}
           cache: "pip"
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install tox tox-gh-actions
+              run: |
+                if [ "$RUNNER_OS" = "Linux" ]; then
+                  sudo apt update
+                  sudo apt install -y libmpich-dev zlib1g-dev
+                elif [ "$RUNNER_OS" = "macOS" ]; then
+                  brew install mpich zlib
+                fi
+                python -m pip install --upgrade pip
+                pip install tox tox-gh-actions
       - name: Run linter
         run: tox -e lint
