### 2025-10-26 ed3beb0
add back mpi
