+++ .github/workflows/typecheck.yaml
@@ -30,15 +30,18 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           cache: "pip"
+
       - name: Install dependencies
-              run: |
-                if [ "$RUNNER_OS" = "Linux" ]; then
-                  sudo apt update
-                  sudo apt install -y libmpich-dev zlib1g-dev
-                elif [ "$RUNNER_OS" = "macOS" ]; then
-                  brew install mpich zlib
-                fi
-                python -m pip install --upgrade pip
-                pip install tox tox-gh-actions
+        run: |
+          if [ "$RUNNER_OS" = "Linux" ]; then
+            sudo apt-get update
+            sudo apt-get install -y libmpich-dev zlib1g-dev
+          elif [ "$RUNNER_OS" = "macOS" ]; then
+            brew update
+            brew install mpich zlib
+          fi
+          python -m pip install --upgrade pip
+          pip install tox tox-gh-actions
+
       - name: Run mypy
         run: tox -e mypy
