+++ .github/workflows/Coverage.yaml
@@ -30,18 +30,22 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           cache: "pip"
+
       - name: Install dependencies
-              run: |
-                if [ "$RUNNER_OS" = "Linux" ]; then
-                  sudo apt update
-                  sudo apt install -y libmpich-dev zlib1g-dev
-                elif [ "$RUNNER_OS" = "macOS" ]; then
-                  brew install mpich zlib
-                fi
-                python -m pip install --upgrade pip
-                pip install tox tox-gh-actions
+        run: |
+          if [ "$RUNNER_OS" = "Linux" ]; then
+            sudo apt-get update
+            sudo apt-get install -y libmpich-dev zlib1g-dev
+          elif [ "$RUNNER_OS" = "macOS" ]; then
+            brew update
+            brew install mpich zlib
+          fi
+          python -m pip install --upgrade pip
+          pip install tox tox-gh-actions
+
       - name: Run coverage
         run: tox -e coverage
+
       - name: Upload coverage report
         uses: actions/upload-artifact@v4
         with:
