### 2025-11-10 #8585 @WilkAndy
Hi @zhubonan - just confirming that I am available to review this submission.  A quick scan of the PDF paper suggests I'll be able to comment on coding aspects (i have done a lot of MPI and cython) and physical aspects (i have done a lot porous-material work).  I am familiar with JOSS reviewing.
**BUT** i won't be able to do this review for ~2 weeks.
