### 2025-12-14 #8585 @WilkAndy
I have almost finished my review, @zhubonan .  I thought that JOSS would automatically give me a checklist.

@tmweigand , i had too much fun with your code!  Very well written, and well documented, and a cute contribution, nice!, and thank you for contributing.  I'm happy to recommend publication, when these few items below have been addressed:

1. Problem with references, as mentioned above by EditorialBot
2. Perhaps include a link to `https://tmweigand.github.io/PMMoTo/index.html` in the PDF?  I'll leave it up to you.
3. Perhaps it was just my setup, but i had to: `python -m pip install typing-extensions pytest-cov`.  I wonder whether the docs should mention `sphinx`?
4. In `sphere_pack_psd.py` line 53 and line 70, need to fix `positional argument follows keyword argument` problems.  Eg, `sd = sd`.
5. I couldn't run `test_benchmark_edt.py` (and similar) due to `No module named 'edt'`.  I don't think that's a deal breaker.
6. But what is a deal breaker is I could not get the tests to run.  I suggest you provide instructions on how to run the tests.  (I just used `pytest`.). I received errors like: `<path>/PMMoTo/tests/core/test_utils.py:82: PytestUnknownMarkWarning: Unknown pytest.mark.mpi - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.mpi(min_size=8)`

Finally, nothing to do with this review, but i think that Tony Roberts, https://www.qut.edu.au/about/our-people/academic-profiles/ap.roberts , might be interested in this work.
