### 2025-12-27 e7165d7
adds libdeflate
