+++ src/pmmoto/io/data_read.hpp
@@ -10,6 +10,7 @@
 #include <fcntl.h>
 #include <unistd.h>
 #include <zlib.h>
+#include <dlfcn.h>
 
 #include <algorithm>
 #include <charconv>
@@ -161,21 +162,74 @@ gzip_uncompressed_size_from_footer(const std::string& filename)
     return n;
 }
 
+#ifdef HAVE_LIBDEFLATE
+  #include <libdeflate.h>
+#endif
+
 static inline std::string
 readGzipFile_fast_zlib(const std::string& filename)
 {
+    using clock = std::chrono::high_resolution_clock;
+    auto t0 = clock::now();
+
+    // Read the compressed file into memory
     std::ifstream ifs(filename, std::ios::binary | std::ios::ate);
     if (UNLIKELY(!ifs)) throw std::runtime_error("open gz failed: " + filename);
-
     size_t comp_size = static_cast<size_t>(ifs.tellg());
     ifs.seekg(0);
-
-    std::string comp;
-    comp.resize(comp_size);
+    std::string comp(comp_size, '\0');
     ifs.read(comp.data(), comp_size);
     if (UNLIKELY(!ifs)) throw std::runtime_error("read gz failed: " + filename);
     ifs.close();
 
+    auto t1 = clock::now();
+
+    size_t out_size = gzip_uncompressed_size_from_footer(filename);
+    std::string out(std::max<size_t>(out_size, 1), '\0');
+
+#ifdef HAVE_LIBDEFLATE
+    // Fast path: libdeflate
+    libdeflate_decompressor* dec = libdeflate_alloc_decompressor();
+    if (!dec) throw std::runtime_error("libdeflate_alloc_decompressor failed");
+
+    size_t actual = 0;
+    libdeflate_result res = libdeflate_gzip_decompress(
+        dec,
+        comp.data(), comp.size(),
+        out.data(), out.size(),
+        &actual);
+
+    if (res == LIBDEFLATE_INSUFFICIENT_SPACE) {
+        out.resize(out.size() * 2 + 65536);
+        res = libdeflate_gzip_decompress(
+            dec,
+            comp.data(), comp.size(),
+            out.data(), out.size(),
+            &actual);
+    }
+    libdeflate_free_decompressor(dec);
+
+    if (res != LIBDEFLATE_SUCCESS) {
+        throw std::runtime_error("libdeflate_gzip_decompress failed: " + std::to_string(res));
+    }
+    out.resize(actual);
+
+    auto t2 = clock::now();
+
+    if (std::getenv("PMMOTO_ZLIB_DEBUG")) {
+        auto ms = [](auto a, auto b) { return std::chrono::duration_cast<std::chrono::milliseconds>(b - a).count(); };
+        const char* libpath = "?";
+    #if defined(__APPLE__) || defined(__linux__)
+        Dl_info info{};
+        if (dladdr((void*)&libdeflate_gzip_decompress, &info) && info.dli_fname) libpath = info.dli_fname;
+    #endif
+        std::fprintf(stderr, "[pmmoto.io] gz=%s read=%lldms inflate=%lldms zlib=libdeflate lib=%s\n",
+                     filename.c_str(), (long long)ms(t0, t1), (long long)ms(t1, t2), libpath);
+        std::fflush(stderr);
+    }
+    return out;
+#else
+    // Fallback: zlib inflate
     size_t out_size = gzip_uncompressed_size_from_footer(filename);
     std::string out;
     out.resize(out_size);
@@ -220,6 +274,7 @@ readGzipFile_fast_zlib(const std::string& filename)
     inflateEnd(&strm);
     out.resize(actual);
     return out;
+#endif
 }
 
 // SIMD-accelerated newline finder (AVX2 version)
