### 2025-12-27 a850586
adds libdeflate
