### 2025-12-27 60f1f30
fixes doi and adds link
