+++ paper/paper.bib
@@ -19,7 +19,7 @@
   pages={243--255},
   year={2001},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/S0309-1708(00)00056-7},
+  doi = {10.1016/S0309-1708(00)00056-7},
 }
 
 @article{Schulz_Becker_2007,
@@ -43,7 +43,7 @@
   pages={6825--6849},
   year={2019},
   publisher={Wiley Online Library},
-  doi = {https://doi.org/10.1029/2018WR024586}
+  doi = {10.1029/2018WR024586}
 }
 
 @article{Sarkisov_Bueno_etal_20,
@@ -55,7 +55,7 @@
   pages={9849--9867},
   year={2020},
   publisher={ACS Publications},
-  doi = {https://doi.org/10.1021/acs.chemmater.0c03575}
+  doi = {10.1021/acs.chemmater.0c03575}
 }
 
 @software{Silversmith_Hilei_24,
@@ -92,7 +92,7 @@ volume = {644},
 pages = {120110},
 year = {2022},
 issn = {0376-7388},
-doi = {https://doi.org/10.1016/j.memsci.2021.120110},
+doi = {10.1016/j.memsci.2021.120110},
 url = {https://www.sciencedirect.com/science/article/pii/S0376738821010504},
 author = {Riley Vickers and Timothy M. Weigand and Cass T. Miller and Orlando Coronell},
 keywords = {Molecular dynamics, Polymerization, Hydration methods, Pore morphology, Pore topology, Medial axis, Minkowski functionals},
@@ -106,19 +106,24 @@ volume = {16},
 pages = {100823},
 year = {2021},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2021.100823},
+doi = {10.1016/j.softx.2021.100823},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711021001151},
 author = {Arnout M.P. Boelens and Hamdi A. Tchelepi},
 keywords = {Minkowski functionals, Minkowski functions, Morphological operations, Python},
 abstract = {The Minkowski functionals and functions are a family of morphological measures and can be used to describe both the morphology (shape) and topology (connectedness) of a system. This paper presents the QuantImPy Python package which can compute both the Minkowski functionals and functions. In addition, this package can efficiently perform basic morphological operations and compute their distance maps. QuantImPy is easy to install, well-documented, integrated with existing Python packages, and open source.}
 }
 
-@misc{Weigand_Vickers_etal_25,
-  author       = {Timothy M. Weigand and Riley Vickers and Cass T. Miller and Orlando Coronell},
-  title        = {Temporal Fluctuations of Non-Equilibrium Water Accessible Pathways in Polyamide Reverse Osmosis Active Layers},
-  year         = {2025},
-  month        = jan,
-  doi          = {10.2139/ssrn.5277423},
-  note         = {Preprint},
-  howpublished = {\url{https://doi.org/10.2139/ssrn.5277423}}
+@article{Weigand_Vickers_etal_25,
+title = {Temporal fluctuations of non-equilibrium water accessible pathways in polyamide reverse osmosis active layers},
+journal = {Journal of Membrane Science Letters},
+volume = {5},
+number = {2},
+pages = {100104},
+year = {2025},
+issn = {2772-4212},
+doi = {10.1016/j.memlet.2025.100104},
+url = {https://www.sciencedirect.com/science/article/pii/S2772421225000133},
+author = {Timothy M. Weigand and Riley Vickers and Cass T. Miller and Orlando Coronell},
+keywords = {Reverse osmosis, Molecular dynamics, Morphology and topology, Transport phenomena},
[…]
