### 2025-12-27 510b13a
adds typing_extensions dependency
