+++ requirements.txt
@@ -4,3 +4,4 @@ mpi4py>=4.0.0
 scipy>=1.11
 connected_components_3d>=3.12.4
 matplotlib>=3.0
+typing_extensions>=4.6
