+++ pyproject.toml
@@ -22,7 +22,8 @@ dependencies = [
     "mpi4py>=4.0.0",
     "scipy>=1.11",
     "connected_components_3d>=3.12.4",
-    "matplotlib>=3.0"
+    "matplotlib>=3.0",
+    "typing_extensions>=4.6"
 ]
 requires-python = ">=3.10"
 
@@ -50,7 +51,6 @@ dev = [
     "ruff>=0.11.5",
     "interrogate>=1.5.0",
     # "pytest-mpi @ git+https://github.com/orausch/pytest-mpi.git@master",
-    "typing_extensions>=4.6"
 ]
 
 docs = [
