+++ benchmarks/test_benchmark_edt.py
@@ -1,9 +1,9 @@
 """Benchmark tests for PMMoTo and edt distance transform implementations."""
 
-from pmmoto import domain_generation
-from pmmoto import filters
 import edt
 import pytest
+from pmmoto import domain_generation
+from pmmoto import filters
 
 
 @pytest.mark.benchmark
@@ -17,7 +17,7 @@ def test_pmmoto_edt(benchmark):
     voxels = (300, 300, 300)
     prob_zero = 0.1
     seed = 1
-    img = domain_generation.gen_random_binary_grid(voxels, prob_zero, seed)
+    img = domain_generation.gen_img_random_binary(voxels, prob_zero, seed)
     _ = benchmark(filters.distance.edt3d, img, periodic=[False, False, False])
 
 
@@ -31,7 +31,7 @@ def test_pmmoto_periodic_edt(benchmark):
     voxels = (300, 300, 300)
     prob_zero = 0.1
     seed = 1
-    img = domain_generation.gen_random_binary_grid(voxels, prob_zero, seed)
+    img = domain_generation.gen_img_random_binary(voxels, prob_zero, seed)
     _ = benchmark(filters.distance.edt3d, img, periodic=[True, True, True])
 
 
@@ -45,5 +45,5 @@ def test_edt(benchmark):
     voxels = (300, 300, 300)
     prob_zero = 0.1
     seed = 1
-    img = domain_generation.gen_random_binary_grid(voxels, prob_zero, seed)
+    img = domain_generation.gen_img_random_binary(voxels, prob_zero, seed)
     _ = benchmark(edt.edt, img)
