+++ benchmarks/test_benchmark_connected_components.py
@@ -16,7 +16,7 @@ def test_pmmoto_connected_components(benchmark):
     prob_zero = 0.1
     seed = 1
     sd = initialize(voxels)
-    img = domain_generation.gen_random_binary_grid(voxels, prob_zero, seed)
+    img = domain_generation.gen_img_random_binary(voxels, prob_zero, seed)
     _ = benchmark(filters.connected_components.connect_components, img, sd)
 
 
@@ -32,5 +32,5 @@ def test_pmmoto_connected_components_periodic(benchmark):
     seed = 1
     boundary_types = ((2, 2), (2, 2), (2, 2))
     sd = initialize(voxels, boundary_types)
-    img = domain_generation.gen_random_binary_grid(sd.voxels, prob_zero, seed)
+    img = domain_generation.gen_img_random_binary(sd.voxels, prob_zero, seed)
     _ = benchmark(filters.connected_components.connect_components, img, sd)
