+++ pyproject.toml
@@ -50,6 +50,7 @@ dev = [
     "pytest_benchmark>=5.1.0",
     "ruff>=0.11.5",
     "interrogate>=1.5.0",
+    "edt>=3.1.0",
     # "pytest-mpi @ git+https://github.com/orausch/pytest-mpi.git@master",
 ]
 
