### 2025-12-27 f7335b9
uses own fork of pytest-mpi
