+++ pyproject.toml
@@ -51,7 +51,7 @@ dev = [
     "ruff>=0.11.5",
     "interrogate>=1.5.0",
     "edt>=3.1.0",
-    # "pytest-mpi @ git+https://github.com/orausch/pytest-mpi.git@master",
+    "pytest-mpi-tmweigand==0.7.0"
 ]
 
 docs = [
