### 2025-12-30 1c7eb2d
adds benchmark fixtures and group names
