+++ benchmarks/test_benchmark_porosimetry_random.py
@@ -3,6 +3,8 @@
 Includes tests for random and inkbottle domains using morph, distance, and hybrid modes.
 """
 
+import pytest
+
 from pmmoto import domain_generation
 from pmmoto import filters
 from pmmoto import initialize
@@ -19,6 +21,7 @@ def setup_random():
     return sd, pm
 
 
+@pytest.mark.benchmark(group="porosimetry_random_small_radius")
 def test_random_morph_mode_small_r(benchmark):
     """Benchmark porosimetry (morph mode, small radius) on random grid."""
     radius = 0.004
@@ -32,6 +35,7 @@ def test_random_morph_mode_small_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_random_large_radius")
 def test_random_morph_mode_large_r(benchmark):
     """Benchmark porosimetry (morph mode, large radius) on random grid."""
     radius = 0.1
@@ -45,6 +49,7 @@ def test_random_morph_mode_large_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_random_small_radius")
 def test_random_distance_small_r(benchmark):
     """Benchmark porosimetry (distance mode, small radius) on random grid."""
     radius = 0.004
@@ -58,6 +63,7 @@ def test_random_distance_small_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_random_large_radius")
 def test_random_distance_large_r(benchmark):
     """Benchmark porosimetry (distance mode, large radius) on random grid."""
     radius = 0.1
@@ -71,6 +77,7 @@ def test_random_distance_large_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_random_small_radius")
 def test_random_hybrid_small_r(benchmark):
     """Benchmark porosimetry (hybrid mode, small radius) on random grid."""
     radius = 0.004
@@ -84,6 +91,7 @@ def test_random_hybrid_small_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_random_large_radius")
 def test_random_hybrid_large_r(benchmark):
     """Benchmark porosimetry (hybrid mode, large radius) on random grid."""
     radius = 0.1
