+++ benchmarks/test_benchmark_connected_components.py
@@ -1,10 +1,30 @@
 """Benchmark tests for PMMoTo connected components filter."""
 
+import pytest
+import cc3d
+import numpy as np
+
 from pmmoto import initialize
 from pmmoto import domain_generation
 from pmmoto import filters
 
 
+@pytest.mark.benchmark(group="connected_components")
+def test_connected_components(benchmark):
+    """Benchmark the connected components filter on a random binary grid.
+
+    Args:
+        benchmark: pytest-benchmark fixture.
+
+    """
+    voxels = (300, 300, 300)
+    prob_zero = 0.1
+    seed = 1
+    img = domain_generation.gen_img_random_binary(voxels, prob_zero, seed)
+    _ = benchmark(cc3d.connected_components, img, return_N=True, out_dtype=np.uint64)
+
+
+@pytest.mark.benchmark(group="connected_components")
 def test_pmmoto_connected_components(benchmark):
     """Benchmark the connected components filter on a random binary grid.
 
@@ -20,6 +40,7 @@ def test_pmmoto_connected_components(benchmark):
     _ = benchmark(filters.connected_components.connect_components, img, sd)
 
 
+@pytest.mark.benchmark(group="connected_components")
 def test_pmmoto_connected_components_periodic(benchmark):
     """Benchmark the connected components filter with periodic boundary conditions.
 
