+++ benchmarks/test_benchmark_porosimetry_inkbottle.py
@@ -3,6 +3,8 @@
 Includes tests for random and inkbottle domains using morph, distance, and hybrid modes.
 """
 
+import pytest
+
 from pmmoto import domain_generation
 from pmmoto import filters
 from pmmoto import initialize
@@ -18,6 +20,7 @@ def setup_inkbottle():
     return sd, pm
 
 
+@pytest.mark.benchmark(group="porosimetry_inkbottle_small_radius")
 def test_inkbottle_morph_mode_small_r(benchmark):
     """Benchmark porosimetry (morph mode, small radius) on inkbottle case."""
     radius = 0.26
@@ -32,6 +35,7 @@ def test_inkbottle_morph_mode_small_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_inkbottle_large_radius")
 def test_inkbottle_morph_mode_large_r(benchmark):
     """Benchmark porosimetry (morph mode, large radius) on inkbottle case."""
     radius = 1.2545
@@ -46,6 +50,7 @@ def test_inkbottle_morph_mode_large_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_inkbottle_small_radius")
 def test_inkbottle_distance_small_r(benchmark):
     """Benchmark porosimetry (distance mode, small radius) on inkbottle case."""
     radius = 0.26
@@ -60,6 +65,7 @@ def test_inkbottle_distance_small_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_inkbottle_large_radius")
 def test_inkbottle_distance_large_r(benchmark):
     """Benchmark porosimetry (distance mode, large radius) on inkbottle case."""
     radius = 1.2545
@@ -74,6 +80,7 @@ def test_inkbottle_distance_large_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_inkbottle_small_radius")
 def test_inkbottle_hybrid_small_r(benchmark):
     """Benchmark porosimetry (hybrid mode, small radius) on inkbottle case."""
     radius = 0.26
@@ -88,6 +95,7 @@ def test_inkbottle_hybrid_small_r(benchmark):
     )
 
 
+@pytest.mark.benchmark(group="porosimetry_inkbottle_large_radius")
 def test_inkbottle_hybrid_large_r(benchmark):
     """Benchmark porosimetry (hybrid mode, large radius) on inkbottle case."""
     radius = 1.2545
