### 2025-12-30 89f7b01
updates docs and adds developer guide
