### 2025-12-30 3c8e9e5
updates workflows so dont run twice.
