+++ .github/workflows/tests.yml
@@ -1,8 +1,13 @@
 name: Tests
 
 on:
-  push:
   pull_request:
+    branches:
+      - "**"
+
+  push:
+    branches:
+      - main
 
 jobs:
   test:
