+++ .github/workflows/typecheck.yaml
@@ -1,8 +1,13 @@
 name: Type Check
 
 on:
-  - push
-  - pull_request
+  pull_request:
+    branches:
+      - "**"
+
+  push:
+    branches:
+      - main
 
 jobs:
   mypy:
