### 2025-12-30 8937cf4
Bump version to 1.0.2
