### 2025-12-30 3b4d0a6
updates runners to osx-latest and macos-15-intel for intel pypi release
