### 2025-12-31 c1445b9
Bump version to 1.0.3
