### 2025-12-31 a563b31
Bump version to 1.0.4
