+++ setup.py
@@ -90,6 +90,7 @@ ext_modules = [
 ]
 setup(
     name="pmmoto",
+    version="1.0.4",
     ext_modules=cythonize(
         ext_modules,
         annotate=True,
