### 2026-01-01 54ab8cb
Bump version to 1.0.5
