+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pmmoto"
-version = "1.0.4"
+version = "1.0.5"
 description = "Porous Media Topology and Morphology Toolkit"
 
 license = { text = "MIT" }
@@ -34,7 +34,13 @@ requires-python = ">=3.10"
 "Issue Tracker" = "https://github.com/tmweigand/PMMoTo/issues"
 
 [build-system]
-requires = ["setuptools", "numpy", "cython"]
+requires = [
+    "setuptools>=64",
+    "wheel",
+    "numpy<2.0",
+    "cython == 3.0"
+]
+
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
