### 2026-01-01 a56ec71
Bump version to 1.0.6
