+++ setup.cfg
@@ -1,0 +1,12 @@
+[metadata]
+name = pmmoto
+version = 1.0.6
+
+[options]
+package_dir =
+    = src
+packages = find:
+zip_safe = False
+
+[options.packages.find]
+where = src
