+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pmmoto"
-version = "1.0.5"
+version = "1.0.6"
 description = "Porous Media Topology and Morphology Toolkit"
 
 license = { text = "MIT" }
