+++ _sources/examples/connected_pathways.rst.txt
@@ -1,0 +1,208 @@
+Connected Pathways in Random Porous Media
+=========================================
+
+This example demonstrates how to identify connected pathways and isolated pores through a **connected components** analysis. The algorithm in PMMoTo is a extension of the Python package ``cc3d`` :cite:`Silversmith_21` for distributed memory systems and periodic porous structures. 
+
+To run this example:
+
+.. code-block:: bash
+
+   mpirun -np 4 python examples/connected_pathways/connected_pathways.py
+
+Step 1: Import Modules
+----------------------
+
+.. code-block:: python
+
+   from mpi4py import MPI
+   import numpy as np
+   import pmmoto
+
+
+Step 2: Domain Setup
+--------------------
+To initialize PMMoTo, the following parameters must be specified for this example:
+
+- ``voxels``: the number of voxels in each dimension for the image representing the porous structure
+- ``boundary_types``: the boundary conditions applied during the simulation
+- ``rank``: the MPI rank of the current process
+- ``box``: the physical size of the simulation domain
+- ``inlet`` and ``outlet``: the designated inflow and outflow faces
+- ``subdomains``: the number of subdomains in each dimension
+
+For this example, we will analyze a porous structure that is thin in the z-dimension to allow for easier visualization. 
+
+.. code-block:: python
+
+   import pmmoto
+   voxels = (2000, 2000, 10)
+   boundary = pmmoto.BoundaryType.END
+   boundary_types = (
+        (boundary, boundary),
+        (boundary, boundary),
+        (boundary, boundary),
+   )
+   box = ((0, 2000), (0, 2000), (0, 30))
+   inlet = ((True, False), (False, False), (False, False))
+   outlet = ((False, True), (False, False), (False, False))
+
+   subdomains = (2, 2, 1)
+
+   sd = pmmoto.initialize(
+        voxels, boundary_types=boundary_types, rank=rank, subdomains=subdomains, box=box, inlet=inlet, outlet=outlet
+   )
+
+The domain decomposed into four subdomains is shown below. 
+
+.. image:: /_static/examples/connected_pathways/subdomains.png
+   :alt: Domain
+   :class: only-light
+   :align: center
+   :width: 60%
+
+Step 3: Generate Random Porous Media
+------------------------------------
+
+While memory-inefficient and generally discouraged, PMMoTo supports reading or generating an entire image and then transferring ownership to the subdomain (i.e., after domain decomposition). To demonstrate this behavior in this example, each process will generate a full image of the entire domain, which will then be decomposed. For this reason, ``subdomain.domain.voxels`` is passed to the domain generation function instead of ``subdomain.voxels``, which only represents the number of voxels owned by the local process.
+
+The porous structure is generated using the function ``gen_img_smoothed_random_binary``. The total number of domain voxels is specified, and to ensure that an identical image is created on each process, the ``seed`` parameter is set. The ``p_zero`` parameter controls the porosity of the structure and represents the probability of a voxel being a pore. The ``smoothness`` parameter determines the degree of spatial smoothing applied to the image.
+
+.. code-block:: python
+
+   img = pmmoto.domain_generation.gen_img_smoothed_random_binary(
+       sd.domain.voxels, p_zero=0.5, smoothness=10, seed=8
+   )
+
+.. image:: /_static/examples/connected_pathways/pore_space.png
+   :alt: Domain
+   :class: only-light
+   :align: center
+   :width: 60%
+
+Step 4: Domain Decomposition
+----------------------------
+
[…]
