+++ _sources/api/pmmoto.domain_generation.rst.txt
@@ -1,0 +1,50 @@
+pmmoto.domain\_generation package
+=================================
+
+.. automodule:: pmmoto.domain_generation
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Submodules
+----------
+
+pmmoto.domain\_generation.domain\_generation module
+---------------------------------------------------
+
+.. automodule:: pmmoto.domain_generation.domain_generation
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pmmoto.domain\_generation.lattice\_packings module
+--------------------------------------------------
+
+.. automodule:: pmmoto.domain_generation.lattice_packings
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pmmoto.domain\_generation.multiphase module
+-------------------------------------------
+
+.. automodule:: pmmoto.domain_generation.multiphase
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pmmoto.domain\_generation.porousmedia module
+--------------------------------------------
+
+.. automodule:: pmmoto.domain_generation.porousmedia
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pmmoto.domain\_generation.rdf module
+------------------------------------
+
+.. automodule:: pmmoto.domain_generation.rdf
+   :members:
+   :show-inheritance:
+   :undoc-members:
