### 2026-01-06 #9711 @tmweigand
> Just copying my comments from [#8585 (comment)](https://github.com/openjournals/joss-reviews/issues/8585#issuecomment-3650185897) . [@tmweigand](https://github.com/tmweigand) , could you please comment on these items? Eg, i think you've fixed the reference issue.
>
> I have almost finished my review, [@zhubonan](https://github.com/zhubonan) . I thought that JOSS would automatically give me a checklist.
>
> [@tmweigand](https://github.com/tmweigand) , i had too much fun with your code! Very well written, and well documented, and a cute contribution, nice!, and thank you for contributing. I'm happy to recommend publication, when these few items below have been addressed:
>
>     * Problem with references, as mentioned above by EditorialBot
>
>     * Perhaps include a link to https://tmweigand.github.io/PMMoTo/index.html in the PDF? I'll leave it up to you.
>
>     * Perhaps it was just my setup, but i had to: python -m pip install typing-extensions pytest-cov. I wonder whether the docs should mention sphinx?
>
>     * In sphere_pack_psd.py line 53 and line 70, need to fix positional argument follows keyword argument problems. Eg, sd = sd.
>
>     * I couldn't run test_benchmark_edt.py (and similar) due to No module named 'edt'. I don't think that's a deal breaker.
>
>     * But what is a deal breaker is I could not get the tests to run. I suggest you provide instructions on how to run the tests. (I just used pytest.). I received errors like: /PMMoTo/tests/core/test_utils.py:82: PytestUnknownMarkWarning: Unknown pytest.mark.mpi - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html @pytest.mark.mpi(min_size=8)
>
>
> Finally, nothing to do with this review, but i think that Tony Roberts, https://www.qut.edu.au/about/our-people/academic-profiles/ap.roberts , might be interested in this work.

Hi @WilkAndy. Thank you much for the very useful and helpful comments! I have addressed each of them in https://github.com/tmweigand/PMMoTo/pull/93.
