### 2026-01-06 tmweigand/PMMoTo#95 @saeedmhz
[JOSS Review] Documentation, Paper Authorship, and Visuals
Hello! I am reviewing this submission for JOSS (see review issue [#9711](https://github.com/openjournals/joss-reviews/issues/9711)). I have a few questions and suggestions regarding the paper and general documentation presentation.

**1. Authorship Clarification**
The paper PDF currently lists only one author, but I noticed there are two contributors listed on the GitHub repository. Could you please clarify if the second contributor should be included as an author on the paper, or confirm that the current author list is correct?

**2. Documentation Badge**
I recommend adding a badge to the top of the `README` that links directly to the documentation site to make it easier for users to find.

**3. Visual Showcase (Figure)**
While not mandatory, I suggest adding a curated figure to the paper or the repository `README`. I believe a visual representation would greatly help users understand the tool's capabilities at a glance.

**4. Paper Typo (Quality of Writing)**
The paper is well-structured and clear. However, I found one small typo in the **Acknowledgements** section that should be fixed:
* **Current Text:** *"Portions of **have** this work have been supported..."*
