### 2026-01-06 tmweigand/PMMoTo#96 @saeedmhz
[JOSS Review] Installation, Testing, and Code Coverage
Hello!

Continuing my review for JOSS (see review issue [#9711](https://github.com/openjournals/joss-reviews/issues/9711)). I have successfully installed the software and run all examples and tests. I have a few technical suggestions regarding the installation and testing process.

### 1. Paraview Installation Recommendations
I attempted to install Paraview using `apt install`. This method resulted in an incomplete installation that did not work with PMMoTo. I found that downloading the executable version of the most recent update fixed this issue.
* **Suggestion:** It would be helpful to update the documentation or `README` to explicitly recommend installing Paraview from the official source. Alternatively, adding a "Troubleshooting" section with this information would be great.

### 2. Developer Mode for Testing
I noticed that in order to run the tests successfully, the package seems to require installation in **developer mode** (`pip install -e .[dev]`).
* **Suggestion:** Could you please add a note to the `README` or the documentation clarifying that an editable install (with dev dependencies) is a prerequisite for running `pytest`?

### 3. Test Coverage
I was able to generate a coverage report, but I noticed that some of the core functionality files currently show around 50% or less coverage.
* Is there a plan to increase this coverage in the near future?
* If increasing coverage is difficult due to the nature of the algorithms, could you briefly explain this context?

Thanks!
