+++ requirements.txt
@@ -1,5 +1,4 @@
-numpy<2.0
-cython==3.0
+numpy>=2.0
 mpi4py>=4.0.0
 scipy>=1.11
 connected_components_3d>=3.12.4
