+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 3.0.0
-envlist = py310, py311, py312, lint, doccheck, coverage, mypy, black
+envlist = py310, py311, py312, py313, lint, doccheck, coverage, mypy, black
 isolated_build = true
 
 [gh-actions]
@@ -8,6 +8,7 @@ python =
     3.10: py310
     3.11: py311
     3.12: py312
+    3.13: py313
 
 [testenv]
 description = Run unit tests with MPI
