### 2026-01-10 7b8ab6a
fix type issues for rdf
