### 2026-01-10 93e1a0f
fix typo in acknowledgements
