### 2026-01-10 352dd98
add image of MD porous structure
