+++ paper/paper.md
@@ -29,7 +29,9 @@ PMMoTo includes parallel implementations of common morphological operations, wit
 
 # Research Application
 
-Recently, PMMoTo was used to generate and analyze highly resolved probabilistic porous structures derived from molecular dynamics (MD) simulations [@Vickers_Weigand_etal_22;@Weigand_Vickers_etal_25] using a voxel resolution of 0.05 Å to produce a domain with 49.5 billion voxels. These simulations were conducted to investigate water transport through reverse osmosis polyamide membranes and assess assumptions regarding the nature of the polyamide structure. With support from observed simulation data, ephemeral water-accessible pathways (i.e. transiently connected pathways that span form the inlet to the outlet of the membrane) were identified. While it is well understood that polyamide membranes are water permeable, PMMoTo enables a novel computational approach that links atomistic simulations to larger scale observations through morphological and topological analysis of the porous structure. This approach aims to improve our understanding of transport across spatial scales and to support the design of improved membranes.
+Recently, PMMoTo was used to generate and analyze highly resolved probabilistic porous structures derived from molecular dynamics (MD) simulations [@Vickers_Weigand_etal_22;@Weigand_Vickers_etal_25] using a voxel resolution of 0.05 Å to produce a domain with 49.5 billion voxels (\autoref{fig:1}). These simulations were conducted to investigate water transport through reverse osmosis polyamide membranes and assess assumptions regarding the nature of the polyamide structure. With support from observed simulation data, ephemeral water-accessible pathways (i.e. transiently connected pathways that span form the inlet to the outlet of the membrane) were identified. While it is well understood that polyamide membranes are water permeable, PMMoTo enables a novel computational approach that links atomistic simulations to larger scale observations through morphological and topological analysis of the porous structure. This approach aims to improve our understanding of transport across spatial scales and to support the design of improved membranes.
+
+![Comparison of porous structures under different assumptions about atom size for simulated polyamide membranes. PMMoTo provides tools to generate and analyze such porous structures. [@Weigand_Vickers_etal_25]. \label{fig:1}](images/comparison_pore_space.png)
 
 # Acknowledgements
 
