+++ _sources/api/pmmoto.io.rst.txt
@@ -1,0 +1,42 @@
+pmmoto.io package
+=================
+
+.. automodule:: pmmoto.io
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pmmoto.io.evtk
+
+Submodules
+----------
+
+pmmoto.io.data\_read module
+---------------------------
+
+.. automodule:: pmmoto.io.data_read
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pmmoto.io.io\_utils module
+--------------------------
+
+.. automodule:: pmmoto.io.io_utils
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pmmoto.io.output module
+-----------------------
+
+.. automodule:: pmmoto.io.output
+   :members:
+   :show-inheritance:
+   :undoc-members:
