+++ _sources/developer_guide/index.rst.txt
@@ -1,0 +1,22 @@
+Developer Guide
+====================
+
+This section is intended for developers and contributors who want to work on PMMoTo itself.
+It covers setting up a development environment, running tests and benchmarks, contributing
+code, and preparing releases. These pages assume some familiarity with Python development
+tools and are not required for typical users of the library.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Developer Guide
+
+   installation
+   python_versions
+   testing
+   tox
+   docs
+   contributing
+   release
+   docker
+   code_of_conduct
+   benchmarking
