+++ _sources/api/pmmoto.filters.distance.rst.txt
@@ -1,0 +1,18 @@
+pmmoto.filters.distance package
+===============================
+
+.. automodule:: pmmoto.filters.distance
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Submodules
+----------
+
+pmmoto.filters.distance.edt module
+----------------------------------
+
+.. automodule:: pmmoto.filters.distance.edt
+   :members:
+   :show-inheritance:
+   :undoc-members:
