+++ _sources/api/pmmoto.rst.txt
@@ -1,0 +1,20 @@
+pmmoto package
+==============
+
+.. automodule:: pmmoto
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pmmoto.analysis
+   pmmoto.core
+   pmmoto.domain_generation
+   pmmoto.filters
+   pmmoto.io
+   pmmoto.particles
