### 2026-01-10 c5300ec
adds badges to readme
