+++ README.md
@@ -1,6 +1,8 @@
 # PMMoTo
 
-![Tests](https://github.com/tmweigand/PMMoTo/actions/workflows/tests.yml/badge.svg)
+[![Documentation](https://img.shields.io/badge/Documentation-PMMoTo-blue?style=for-the-badge)](https://tmweigand.github.io/PMMoTo/)
+
+![Tests](https://github.com/tmweigand/PMMoTo/actions/workflows/tests.yml/badge.svg) [![Python](https://img.shields.io/badge/python-3.10%20%7C%203.11%20%7C%203.12%20%7C%203.13-blue?style=flat-square)](https://www.python.org/) [![PyPI](https://img.shields.io/pypi/v/pmmoto?style=flat-square)](https://pypi.org/project/pmmoto/) [![License](https://img.shields.io/badge/License-MIT-blue?style=flat-square)](LICENSE)
 
 ![logo-1-text](https://github.com/tmweigand/PMMoTo/assets/68024672/5f667c8f-5498-4597-9af0-76fd6a9bc59a)
 
@@ -12,7 +14,7 @@ The Porous Media Morphology and Topology (PMMoTo) toolkit is an open-source Pyth
 
 ## Website
 
-tmweigand.github.io/PMMoTo/
+[PMMoTo Website: https://tmweigand.github.io/PMMoTo/](https://tmweigand.github.io/PMMoTo/)
 
 ## Installation
 
