### 2026-01-11 56a4235
adds more testing and fixes morph bug when resolution is close to radius
