### 2026-01-12 ad59030
adds more tests for connected components analysis
