### 2026-01-12 eca647f
tests for lattice packings
