### 2026-01-12 01ec287
adds more tests for particles and fixes periodic addition issue by including boundary types
