+++ tests/particles/test_particles.py
@@ -1,8 +1,9 @@
 """test_particles.py"""
 
+import pmmoto
+import pytest
 import numpy as np
 from mpi4py import MPI
-import pmmoto
 
 
 def test_particles():
@@ -159,8 +160,6 @@ def test_spheres():
 
     spheres.build_KDtree()
 
-    print(spheres.return_np_array())
-
 
 def test_cleanup():
     """Test deletion of particle lists"""
@@ -277,3 +276,193 @@ def test_count_own():
 
     assert atoms.size() == {1: 1, 3: 1, 15: 3}
     assert atoms.get_own_count() == {1: 1, 3: 1, 15: 2}
+
+
+def test_elements_to_ids():
+    """Tests for conversion"""
+    atom_elements = ["H", "C", "O", "N", "Cl"]
+
+    atom_ids = pmmoto.particles.convert_atoms_elements_to_ids(atom_elements)
+
+    np.testing.assert_array_equal(atom_ids, np.array([1, 6, 8, 7, 17]))
+
+
+def test_load_uff_data_from_custom_file(tmp_path):
+    """Test _load_uff_data with a custom UFF file"""
+    # Create a custom UFF file
+    uff_file = tmp_path / "custom_uff.txt"
+    uff_content = """AtomicNumber Name Diameter
+1 H 2.886
+6 C 3.851
+8 O 3.500
+7 N 3.660
+17 Cl
+"""
+    uff_file.write_text(uff_content)
+
+    # Load data from custom file
+    element_table = pmmoto.particles.particles._load_uff_data(str(uff_file))
+
+    # Test lookup by name
+    assert "H" in element_table
+    assert "C" in element_table
+    assert "O" in element_table
+    assert "N" in element_table
+    assert "Cl" not in element_table
+
+    # Test lookup by atomic number
+    assert 1 in element_table
+    assert 6 in element_table
+    assert 8 in element_table
+    assert 7 in element_table
+
+    # Test values (diameter should be halved to radius)
+    atomic_num_h, radius_h = element_table["H"]
+    assert atomic_num_h == 1
+    assert radius_h == pytest.approx(2.886 / 2.0)
+
+    atomic_num_c, radius_c = element_table[6]
+    assert atomic_num_c == 6
+    assert radius_c == pytest.approx(3.851 / 2.0)
+
+
+def test_uff_radius_inputs():
+    """Ensures correct behavior for inputs"""
+    atom_names = ["C", "H"]
+    atomic_numbers = [6, 1]
+
+    with pytest.raises(ValueError):
+        _ = pmmoto.particles.uff_radius(atom_names, atomic_numbers)
+
+
+def test_atom_initialization_order():
+    """Test initialization of atoms"""
+    sd = pmmoto.initialize((10, 10, 10))
+    atom_coordinates = np.array([[0.0, 0.0, 0.0], [0, 1.0, 1.0], [0.5, 0.5, 0.5]])
+    atom_coordinates_f = np.array(atom_coordinates, order="F")
+
+    atom_radii = {1: 0.2, 2: 0.1, 3: 0.05}
+    atom_ids = np.array([1, 2, 3])
+
+    assert not atom_coordinates_f.flags["C_CONTIGUOUS"]
+
+    particles_c = pmmoto.particles.initialize_atoms(
+        sd, atom_coordinates, atom_radii, atom_ids
+    )
+
+    particles_f = pmmoto.particles.initialize_atoms(
+        sd, atom_coordinates_f, atom_radii, atom_ids
+    )
+
+    np.testing.assert_array_equal(
[…]
