### 2026-01-12 2aa4b4c
adds test for io_utils and modifes exit from failure
