### 2026-01-12 21dc57b
adds tests for bins
