+++ tests/io/test_output.py
@@ -1,0 +1,167 @@
+"""test_output.py"""
+
+import pmmoto
+import pytest
+import numpy as np
+from mpi4py import MPI
+
+
+def test_save_particle_data(tmp_path):
+    """Tests for saving particle data"""
+    sd = pmmoto.initialize((10, 10, 10))
+    particles = np.array([[0, 0, 0], [1, 1, 1]])
+    file_path = tmp_path / "particles"
+
+    pmmoto.io.output.save_particle_data(str(file_path), sd, particles)
+    expected_file = file_path / "particlesProc.0.vtu"
+
+    assert expected_file.exists()
+
+    # Check for ptvu file
+    expected_file = file_path / "particles.pvtu"
+    assert expected_file.exists()
+
+    sd.rank = 1
+    particles = np.array([[0, 0, 0, 1], [1, 1, 1, 1]])
+    file_path = tmp_path / "particles"
+
+    pmmoto.io.output.save_particle_data(str(file_path), sd, particles)
+    expected_file = file_path / "particlesProc.1.vtu"
+
+    assert expected_file.exists()
+
+    sd.rank = 2
+    particles = np.array([[0, 0, 0, 1, 2], [1, 1, 1, 1, 2]])
+    file_path = tmp_path / "particles"
+
+    pmmoto.io.output.save_particle_data(str(file_path), sd, particles)
+    expected_file = file_path / "particlesProc.2.vtu"
+
+    assert expected_file.exists()
+
+    sd.rank = 3
+    particles = np.array([[0, 0, 0, 1, 2, 3], [1, 1, 1, 1, 2, 3]])
+    file_path = tmp_path / "particles"
+
+    pmmoto.io.output.save_particle_data(str(file_path), sd, particles)
+    expected_file = file_path / "particlesProc.3.vtu"
+
+    assert expected_file.exists()
+
+
+def test_save_img(tmp_path):
+    """Tests saving of img"""
+    sd = pmmoto.initialize((10, 10, 10))
+    img = np.zeros(sd.voxels)
+    file_path = tmp_path / "test_img"
+
+    pmmoto.io.output.save_img(str(file_path), sd, img)
+
+    expected_file = tmp_path / "test_img.vti"
+    assert expected_file.exists()
+
+    # Img is not shape of sd.voxels
+    err_img = np.zeros((5, 5, 5))
+    with pytest.raises(ValueError):
+        pmmoto.io.output.save_img(
+            str(file_path), sd, img, additional_img={"err_img": err_img}
+        )
+
+
+def test_save_img_additional(tmp_path):
+    """Tests saving of img"""
+    sd = pmmoto.initialize((10, 10, 10))
+    img = np.zeros(sd.voxels)
+    img2 = np.zeros(sd.voxels)
+    file_path = tmp_path / "test_img"
+
+    pmmoto.io.output.save_img(str(file_path), sd, img, additional_img={"img2": img2})
+
+    expected_file = tmp_path / "test_img.vti"
+    assert expected_file.exists()
+
+    pmmoto.io.output.save_img(str(file_path), sd, img, additional_img={"img2": img2})
+
+    expected_file = tmp_path / "test_img.vti"
+    assert expected_file.exists()
+
+    with pytest.raises(TypeError):
+        pmmoto.io.output.save_img(str(file_path), sd, img, additional_img={"img2": 4.0})
+
+    img3 = np.zeros((2, 2, 2))
+    with pytest.raises(ValueError):
+        pmmoto.io.output.save_img(
+            str(file_path), sd, img, additional_img={"img3": img3}
+        )
+
+
[…]
