### 2026-01-14 addefcd
complete tests for subdomain_features
