### 2026-01-14 bdfbb97
complete tests for subdomains
