### 2026-01-14 a8576a9
full tests for padded subdomains
