### 2026-01-14 c3b34ad
more tests for morphological operators
