### 2026-01-14 c58d092
tests for reading of data
