+++ tests/core/test_voxels.py
@@ -1,9 +1,9 @@
 """test_voxels.py"""
 
-import numpy as np
 import pmmoto
 from pmmoto.core import _voxels
 import pytest
+import numpy as np
 
 
 def test_voxls_get_id() -> None:
@@ -344,3 +344,266 @@ def test_get_nearest_boundary_index_figure(generate_padded_subdomain) -> None:
                         img_out[nx, ny, int(index[nx, ny])] = 0
                     else:
                         img_out[nx, ny, 0] = 255
+
+
+def test_get_nearest_boundary_index_errors():
+    """Test for ensuring get_nearest_boundary_index works"""
+    sd = pmmoto.initialize((10, 10, 10))
+    img = pmmoto.domain_generation.gen_img_smoothed_random_binary(sd.voxels, 0.5, 5.0)
+    with pytest.raises(ValueError):
+        _ = pmmoto.core.voxels.get_nearest_boundary_index(
+            subdomain=sd, img=img, label=0, dimension=5
+        )
+
+    sd = pmmoto.initialize((10, 10, 10), return_subdomain=True)
+    with pytest.raises(TypeError):
+        _ = pmmoto.core.voxels.get_nearest_boundary_index(
+            subdomain=sd, img=img, label=0, which_voxels="own"
+        )
+
+
+def test_get_nearest_boundary_index_all() -> None:
+    """Test for ensuring get_nearest_boundary_index works"""
+    sd = pmmoto.initialize((10, 10, 10))
+    img = np.ones(sd.voxels, dtype=np.uint8)
+    img[0, :, :] = 0
+    img[:, 0, :] = 0
+    img[:, :, 0] = 0
+    img[-1, :, :] = 0
+    img[:, -1, :] = 0
+    img[:, :, -1] = 0
+    boundary_index = pmmoto.core.voxels.get_nearest_boundary_index(
+        subdomain=sd, img=img, label=0, which_voxels="all"
+    )
+
+    np.testing.assert_array_equal(boundary_index[(-1, 0, 0)], np.zeros([10, 10]))
+    np.testing.assert_array_equal(boundary_index[(0, -1, 0)], np.zeros([10, 10]))
+    np.testing.assert_array_equal(boundary_index[(0, 0, -1)], np.zeros([10, 10]))
+
+    np.testing.assert_array_equal(boundary_index[(1, 0, 0)], 9 * np.ones([10, 10]))
+    np.testing.assert_array_equal(boundary_index[(0, 1, 0)], 9 * np.ones([10, 10]))
+    np.testing.assert_array_equal(boundary_index[(0, 0, 1)], 9 * np.ones([10, 10]))
+
+
+def test_get_nearest_boundary_index_own() -> None:
+    """Test for ensuring get_nearest_boundary_index works"""
+    sd = pmmoto.initialize((10, 10, 10))
+    img = np.ones(sd.voxels, dtype=np.uint8)
+    img[0, :, :] = 0
+    img[:, 0, :] = 0
+    img[:, :, 0] = 0
+    img[-1, :, :] = 0
+    img[:, -1, :] = 0
+    img[:, :, -1] = 0
+    boundary_index = pmmoto.core.voxels.get_nearest_boundary_index(
+        subdomain=sd, img=img, label=0, which_voxels="own"
+    )
+
+    np.testing.assert_array_equal(boundary_index[(-1, 0, 0)], np.zeros([10, 10]))
+    np.testing.assert_array_equal(boundary_index[(0, -1, 0)], np.zeros([10, 10]))
+    np.testing.assert_array_equal(boundary_index[(0, 0, -1)], np.zeros([10, 10]))
+
+    np.testing.assert_array_equal(boundary_index[(1, 0, 0)], 9 * np.ones([10, 10]))
[…]
