### 2026-01-15 d5f6a0f
more tests for communation and better error handling
