### 2026-01-15 e3b88f0
complete tests for domain generation
