### 2026-01-15 300fcd0
adds test for domains
