### 2026-01-15 5a353eb
fixes bad tests and adds test for utils
