### 2026-01-15 aa03510
adds test for voxels
