### 2026-01-16 7d07d08
type fixes and logger test
