+++ _sources/examples/md_porous_media.rst.txt
@@ -1,0 +1,212 @@
+Porous Structure from Molecular Dynamics Simulations
+======================================================
+
+This example demonstrates how to generate a porous structure from molecular dynamics (MD) simulations using two approaches to determine atomic radii:
+
+1. **Non-equilibrium radii**, derived from the radial distribution function (RDF) computed during non-equilibrium MD simulations.
+2. **Equilibrium radii** estimated from the Columbic radius of water by the Universal Force Field (UFF).
+
+This example uses MD simulation data :cite:`Vickers_Weigand_etal_22` of water permeation through a polyamide reverse osmosis membrane that consists of M-phenylenediamine (MPD), trimesoyl chloride (TMC), and water. 
+
+To run:
+
+.. code-block:: bash
+
+   mpirun -np 8 python examples/md_porous_media/md_porous_media.py
+
+
+Step 1: Import Modules
+-------------------------
+
+.. code-block:: python
+
+   from mpi4py import MPI
+   import pmmoto
+
+
+Step 2: Initialize MD Domain
+------------------------------
+
+The input parameters for PMMoTo match the MD simulation parameters. To initialize PMMoTo, the following parameters must be specified for this example:
+
+- ``voxels``: the number of voxels in each dimension for the image representing the porous structure
+- ``box``: the physical size of the simulation domain
+- ``rank``: the MPI rank of the current process
+- ``subdomains``: the number of subdomains in each dimension
+- ``boundary_types``: the boundary conditions applied during the simulation
+- ``verlet_domains``: the number of Verlet domains for each ``Subdomain``
+- ``inlet`` and ``outlet``: the designated inflow and outflow faces
+
+.. note::
+
+   ``verlet_domains`` specify how to subdivide a subdomain into smaller regions to optimize domain generation by reducing the number of particles or atoms that must be checked during computations. Conducting a numerical study to determine the optimal number of Verlet domains can significantly reduce compute times for systems with a large number of atoms. Note that the optimal value depends on the hardware and computational environment.
+
+
+
+.. code-block:: python
+
+   def initialize_domain(voxels: tuple[int, ...]):
+      """Initialize the membrane domain"""
+      subdomains = (2, 2, 2)
+
+      # Full domain with reservoirs
+      box = [
+         [0.0, 176],
+         [0.0, 176],
+         [-100, 100],
+      ]
+
+      sd = pmmoto.initialize(
+         voxels=voxels,
+         box=box,
+         rank=rank,
+         subdomains=subdomains,
+         boundary_types=(
+               (pmmoto.BoundaryType.PERIODIC, pmmoto.BoundaryType.PERIODIC),
+               (pmmoto.BoundaryType.PERIODIC, pmmoto.BoundaryType.PERIODIC),
+               (pmmoto.BoundaryType.END, pmmoto.BoundaryType.END),
+         ),
+         verlet_domains=(20, 20, 20),
+         inlet=((False, False), (False, False), (True, False)),
+         outlet=((False, False), (False, False), (False, True)),
+      )
+
+      return sd
+
+Step 3: Equilibrium Radii
+---------------------------------------
+
+The **equilibrium approach** for reconstructing a porous structure from MD simulations uses the van der Waals surface, based on element-specific radii defined by the Universal Force Field (UFF) :cite:`Sarkisov_Bueno_etal_20`. In the MD simulations in this work, four elements are needed to represent the MPD and TMC, namely carbon,  nitrogen, oxygen, and hydrogen. 
+
+Additionally, as we are interested in water-accessible pathways, an additional 1.4 Å is added to each atom radius, accounting for the Columbic radius of water. 
+
+``atom_folder`` is a text file with the following format:
+
[…]
