+++ _sources/examples/drainage_inkbottle.rst.txt
@@ -1,0 +1,240 @@
+Drainage of an Ink Bottle
+=========================
+
+This example demonstrates a morphological drainage simulation of an ink bottle using PMMoTo. A morphological drainage simulation uses morphologic methods such as erosion and dilation to approximate equilibrium states of the displacement of a wetting phase by a non-wetting phase in a porous structure. For the simulation, a porous structure and capillary pressure are specified and the equilibrium state (as a multiphase image) and wetting-phase saturation is output.  Two drainage approaches are examined in this example:
+
+- A standard morphological approach :cite:`Hilpert_Miller_2001`
+- An extension to the standard approach where the contact angle can be specified :cite:`Schulz_Becker_2007`
+
+To run this example:
+
+.. code-block:: bash
+
+   mpirun -np 2 python examples/drainage_inkbottle/drainage_inkbottle.py
+
+Step 1: Import Modules
+----------------------
+This script requires ``MPI`` for parallelism, however, ``pmmoto`` handles all communication. ``numpy`` is needed to generate an list of capillary pressures for the simulations and ``matplotlib`` is used for plotting.  
+
+.. code-block:: python
+
+   from mpi4py import MPI
+   import numpy as np
+   import matplotlib.pyplot as plt
+   import pmmoto
+
+Step 2: Initialize PMMoTo
+--------------------------------
+
+To initialize PMMoTo, several parameters must be specified:
+
+- ``voxels``: the number of voxels in each dimension for the image representing the porous structure
+- ``reservoir_voxels``: the number of voxels added to the inlet
+- ``subdomains``: the number of subdomains in each dimension
+- ``boundary_types``: the boundary conditions applied during the simulation
+- ``inlet`` and ``outlet``: the designated inflow and outflow faces
+- ``box``: the physical size of the simulation domain
+- ``rank``: the MPI rank of the current process
+
+These parameters collectively define the system's geometry, boundary conditions, and how it is partitioned across processes for the morphological drainage simulation.
+
+.. code-block:: python
+
+    # Domain voxel resolution.
+    # Must be a 3-tuple as only 3D is currently supported.
+    voxels = (560, 120, 120)
+
+    # Number of voxels allocated for the inlet reservoir region.
+    reservoir_voxels = 20
+
+    # Domain decomposition across MPI ranks.
+    # The product of subdomain counts must match the number of MPI processes.
+    subdomains = (2, 1, 1)
+
+    # Physical extent of the domain in each dimension: (min, max)
+    box = (
+        (0.0, 14.0),  # x-dimension
+        (-1.5, 1.5),  # y-dimension
+        (-1.5, 1.5),  # z-dimension
+    )
+
+    # Boundary conditions for each face (−, +) per axis.
+    # Options:
+    #   - END:     Nothing assumed
+    #   - WALL:    Solid, impermeable boundary added to image
+    #   - PERIODIC: Wraparound — both faces of the axis must be periodic
+    boundary_types = (
+        (pmmoto.BoundaryType.END, pmmoto.BoundaryType.END),  # x
+        (pmmoto.BoundaryType.END, pmmoto.BoundaryType.END),  # y
+        (pmmoto.BoundaryType.END, pmmoto.BoundaryType.END),  # z
+    )
+
+    # Inlet boundary condition: (−, +) per axis
+    # Used to specify where fluid enters the domain
+    inlet = (
+        (False, True),  # x: fluid enters from the +x face
+        (False, False),  # y: no inlet
+        (False, False),  # z: no inlet
+    )
+
+    # Outlet boundary condition: (−, +) per axis
+    # Used to specify where fluid exits the domain
+    outlet = (
[…]
