### 2026-01-16 #9711 @WilkAndy
> [@zhubonan](https://github.com/zhubonan) [@saeedmhz](https://github.com/saeedmhz) [@WilkAndy](https://github.com/WilkAndy) Thanks for the very helpful comments! They have certainly driven improvements to the package. I believe I have now addressed them all.

Your repo is constantly evolving, @tmweigand , which makes it a bit difficult to complete our checklist.  I suggest for the purposes of JOSS that you pause development (or put all such into a different branch) and let's quickly get this JOSS article published referring to a PMMoTo-release in which the tests pass, the doco is adequate for that release, etc.   Without doubt your work is excellent and useful and relevant, and it should be published, but the churn makes it difficult to review.
