### 2026-01-18 7dfdfb0
initialize box and clean up periodic addition of atoms
