+++ src/pmmoto/particles/_particles.pyx
@@ -24,25 +24,32 @@ from ..core import communication
 __all__ = ["_initialize_atoms","_initialize_spheres","_initialize_cylinders"]
 
 
-def create_box(bounds,boundary_type = None):
+def create_box(bounds, boundary_type=None):
     """
-    Convert a tuple of tuples ((min_x, max_x), (min_y, max_y), (min_z, max_z))
-    into a C++ Box structure in Cython.
+    Convert ((min_x, max_x), (min_y, max_y), (min_z, max_z))
+    into a C++ Box structure.
+
+    An axis is periodic only if BOTH lower and upper boundaries
+    are BoundaryType.PERIODIC.
     """
     cdef Box box
-    
-    # Directly assign values from the Python tuple
-    box.min[0], box.max[0] = bounds[0]
-    box.min[1], box.max[1] = bounds[1]
-    box.min[2], box.max[2] = bounds[2]
+    cdef int i
 
+    # Assign bounds and lengths
+    for i in range(3):
+        box.min[i] = bounds[i][0]
+        box.max[i] = bounds[i][1]
+        box.length[i] = box.max[i] - box.min[i]
+        box.periodic[i] = False  # default: non-periodic
+
+    # Set periodic flags based on boundary types
     if boundary_type is not None:
-        box.lower_periodic[0] = 1 if boundary_type[0][0] == boundary_types.BoundaryType.PERIODIC else 0
-        box.upper_periodic[0] = 1 if boundary_type[0][1] == boundary_types.BoundaryType.PERIODIC else 0
-        box.lower_periodic[1] = 1 if boundary_type[1][0] == boundary_types.BoundaryType.PERIODIC else 0
-        box.upper_periodic[1] = 1 if boundary_type[1][1] == boundary_types.BoundaryType.PERIODIC else 0
-        box.lower_periodic[2] = 1 if boundary_type[2][0] == boundary_types.BoundaryType.PERIODIC else 0
-        box.upper_periodic[2] = 1 if boundary_type[2][1] == boundary_types.BoundaryType.PERIODIC else 0
+        for i in range(3):
+            if (
+                boundary_type[i][0] == boundary_types.BoundaryType.PERIODIC
+                and boundary_type[i][1] == boundary_types.BoundaryType.PERIODIC
+            ):
+                box.periodic[i] = True
 
     return box
 
@@ -218,7 +225,7 @@ cdef class PyAtomList:
         cdef Box _domain_box, _subdomain_box
 
         _domain_box = create_box(subdomain.domain.box,subdomain.domain.boundary_types)
-        _subdomain_box = create_box(subdomain.box,subdomain.domain.boundary_types)
+        _subdomain_box = create_box(subdomain.box)
 
         self._atom_list.get().add_periodic_atoms(_domain_box, _subdomain_box)
 
@@ -319,7 +326,7 @@ cdef class PySphereList:
         cdef Box _domain_box,_subdomain_box
 
         _domain_box = create_box(subdomain.domain.box,subdomain.domain.boundary_types)
-        _subdomain_box = create_box(subdomain.box,subdomain.domain.boundary_types)
+        _subdomain_box = create_box(subdomain.box)
 
         self._sphere_list.get().add_periodic_spheres(_domain_box,_subdomain_box)
 
