+++ src/pmmoto/particles/spheres.hpp
@@ -64,10 +64,8 @@ public:
     }
 
     /**
-     * @brief Determines if a particle crosses the outer box boundaries and has
-     * periodic boundary conditions
-     * @param Sphere sphere to check
-     * @param outerBox The outer periodic box dimensions
+     * @brief Determines if a sphere crosses the boundaries of box
+     * @param box Box stuct
      * @return Array of bools indicating whether the particle crosses each axis
      * boundary.
      */
@@ -77,16 +75,34 @@ public:
 
         for (std::size_t d = 0; d < 3; ++d)
         {
-            const bool crosses_lower =
-                box.lower_periodic[d] && (coordinates[d] - radius < box.min[d]);
-
-            const bool crosses_upper =
-                box.upper_periodic[d] && (coordinates[d] + radius > box.max[d]);
-
+            const bool crosses_lower = coordinates[d] - radius < box.min[d];
+            const bool crosses_upper = coordinates[d] + radius > box.max[d];
             crosses[d] = crosses_lower || crosses_upper;
         }
 
         return crosses;
+    }
+
+    /**
+     * @brief Determines whether a sphere extends beyond the box extents.
+     *
+     * This function performs a purely geometric check and does NOT consider
+     * periodicity or boundary conditions.
+     *
+     * @param box Axis-aligned bounding box
+     * @return Array indicating extension beyond box in each dimension
+     */
+    inline std::array<bool, 3> extends_outside_box(const Box& box) const
+    {
+        std::array<bool, 3> outside{};
+
+        for (std::size_t d = 0; d < 3; ++d)
+        {
+            outside[d] = (coordinates[d] - radius < box.min[d]) ||
+                         (coordinates[d] + radius > box.max[d]);
+        }
+
+        return outside;
     }
 
     /**
@@ -346,57 +362,42 @@ public:
      * @param domain Dimensions of domain via Box with periodic boundary info
      * @param subdomain Dimensions of the subdomain vix Box
      */
-    void add_periodic_spheres(Box& domain, const Box& subdomain)
+    void add_periodic_spheres(const Box& domain, const Box& subdomain)
     {
-        domain.box_length();
         std::vector<Sphere> new_spheres;
         std::vector<Coords> new_coords;
+
         static constexpr std::array<int, 3> shifts = { -1, 0, 1 };
-        for (auto& sphere : spheres)
+
+        for (const auto& sphere : spheres)
         {
-            auto crosses = sphere.cross_boundary(domain);
+            const auto extends = sphere.extends_outside_box(domain);
 
             for (int dx : shifts)
             {
-                if (dx != 0 && !crosses[0]) continue;
+                if (dx != 0 && !(extends[0] && domain.periodic[0])) continue;
 
                 for (int dy : shifts)
                 {
-                    if (dy != 0 && !crosses[1]) continue;
+                    if (dy != 0 && !(extends[1] && domain.periodic[1]))
+                        continue;
 
                     for (int dz : shifts)
                     {
-                        if (dz != 0 && !crosses[2]) continue;
-                        if (dx == 0 && dy == 0 && dz == 0)
-                        {
-                            if (sphere.intersects_box(sphere.coordinates,
-                                                      sphere.radius,
-                                                      subdomain))
-                            {
-                                new_coords.push_back(sphere.coordinates);
-                                new_spheres.push_back(sphere);
-                            }
-                            continue; // Skip original sphere
-                        }
+                        if (dz != 0 && !(extends[2] && domain.periodic[2]))
+                            continue;
 
-                        if ((dx != 0 && crosses[0]) ||
-                            (dy != 0 && crosses[1]) || (dz != 0 && crosses[2]))
+                        Coords image = sphere.coordinates;
+
[…]
