### 2026-01-30 #9711 @zhubonan
Hi @tmweigand apologies for the wait. It looks good to me, and I am happy to move to the next stage. I have made some edits to the paper.md through a PR.

The next thing is to create an archive of the software. Please do the following step:

1. Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
2. Archive the release on Zenodo/figshare/etc and post the DOI here.
3. Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
3. Make sure that the license listed for the archive is the same as the software license.

I can see you already have a recent release. You can use Zenodo to archive this release to get an DOI and let me know here.
The idea is that the JOSS publication should be linked with the archive (i.e., this is version of the code that was reviewed).
