### 2026-01-30 50c7798
Bump version to 1.0.7
