+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pmmoto
-version = 1.0.6
+version = 1.0.7
 
 [options]
 package_dir =
