+++ _sources/developer_guide/code_of_conduct.rst.txt
@@ -1,0 +1,32 @@
+Code of Conduct
+===============
+
+PMMoTo is committed to providing a welcoming and respectful community for all contributors, users, and collaborators.  
+
+All participants are expected to:
+
+- Treat everyone with respect, regardless of background, experience, or identity.
+- Communicate constructively and politely.
+- Give and receive feedback gracefully.
+- Report any behavior that violates this code to the project maintainers.
+
+Unacceptable behavior includes, but is not limited to:
+
+- Harassment, intimidation, or discrimination in any form.
+- Offensive comments or language.
+- Public or private harassment of any community member.
+- Disruption of project activities or communities.
+
+**Reporting Issues**
+
+If you experience or witness unacceptable behavior, please contact the project maintainers via email at: ``timothy.weigand@unc.edu``
+
+All reports will be handled with confidentiality and seriousness.  
+
+**Enforcement**
+
+Project maintainers may take any action they deem appropriate, including removal from the community or revocation of contributor privileges, to ensure a safe and welcoming environment.
+
+**Acknowledgment**
+
+By participating in the PMMoTo community, you agree to abide by this Code of Conduct.

[…]
