### 2026-01-30 #9711 @tmweigand
Hi @zhubonan Thanks for the instructions. The release is v1.0.7 and has been released on Zenodo: https://doi.org/10.5281/zenodo.18435235

Also thanks for catching the typos in the paper!
