### 2026-02-03 #9711 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
