### 2026-02-18 #9711 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS (and apologies for the delay in getting this processing started!). Below are some final checks, some of which may require your attention:
As you can see, most seems in order, however the below are some points that require your attention (just two minor copyediting things) :point_down: :
- [x] In Summary, the URL is running over the line, and you don't need to include it anyway since as you can see that link will be there in the sidebar, so please just remove that and leave only the software acronym in the parentheses
- [x] under Research Application: hyphenate "larger-scale"
