### 2026-02-19 04d9a25
removes url and adds hyphens
