+++ paper/paper.md
@@ -19,17 +19,17 @@ bibliography: paper.bib
 
 # Summary
 
-The Porous Media Morphology and Topology Toolkit (PMMoTo, https://tmweigand.github.io/PMMoTo/index.html) is a software tool designed to help researchers analyze and understand porous structures and how their features influence behavior at larger scales. A porous medium is any solid material that contains pores, for example, soil, membranes, skin, and many other natural and engineered materials. In fact, porosity is often a function of spatial scale; all materials have some degree of porosity — you just need to zoom in. The morphology (i.e., the shape and structure) and topology (i.e., the connectivity and spatial relations) of a porous structure govern larger scale behavior. By characterizing these structures, PMMoTo enables researchers to better understand emergent behaviors and develop more predictive models of porous systems.
+The Porous Media Morphology and Topology Toolkit (PMMoTo) is a software tool designed to help researchers analyze and understand porous structures and how their features influence behavior at larger scales. A porous medium is any solid material that contains pores, for example, soil, membranes, skin, and many other natural and engineered materials. In fact, porosity is often a function of spatial scale; all materials have some degree of porosity, you just need to zoom in. The morphology (i.e. the shape and structure) and topology (i.e., the connectivity and spatial relations) of a porous structure govern larger-scale behavior. By characterizing these structures, PMMoTo enables researchers to better understand emergent behaviors and develop more predictive models of porous systems.
 
 # Statement of Need
 
-While several excellent packages, such as PoreSpy [@Gostick_Aghighi_etal_19], support the analysis of porous structures, PMMoTo is distinct in two key regards: (1) PMMoTo focuses on computationally generated, often time-varying, porous structures rather than experimental or image-based data; and (2) PMMoTo is designed to handle large and/or periodic domains. Using computationally generated porous structures avoids the limits of experimental imaging resolution but introduces the need for grid resolution studies and numerical assessment of resolution effects. Highly resolved systems are often required for accurate analysis, making efficient computational methods essential. To ensure scalability and efficiency, performance-critical components of PMMoTo are implemented in C++ and Cython. For many research applications, especially those involving upscaling, large and periodic domains are required to obtain statistically representative estimates of larger scale properties. To handle these requirements, PMMoTo is optimized for distributed-memory systems using the Message Passing Interface (MPI). Additionally, PMMoTo supports the specification of boundary conditions, including periodic boundaries, to match requirements for simulations of porous structures.
+While several excellent packages, such as PoreSpy [@Gostick_Aghighi_etal_19], support the analysis of porous structures, PMMoTo is distinct in two key regards: (1) PMMoTo focuses on computationally generated, often time-varying, porous structures rather than experimental or image-based data; and (2) PMMoTo is designed to handle large and/or periodic domains. Using computationally generated porous structures avoids the limits of experimental imaging resolution but introduces the need for grid resolution studies and numerical assessment of resolution effects. Highly resolved systems are often required for accurate analysis, making efficient computational methods essential. To ensure scalability and efficiency, performance-critical components of PMMoTo are implemented in C++ and Cython. For many research applications, especially those involving upscaling, large and periodic domains are required to obtain statistically representative estimates of larger-scale properties. To handle these requirements, PMMoTo is optimized for distributed-memory systems using the Message Passing Interface (MPI). Additionally, PMMoTo supports the specification of boundary conditions, including periodic boundaries, to match requirements for simulations of porous structures.
 
[…]
