### 2026-02-19 #9711 @tmweigand
Hi @rkurchin! Thanks for catching those. I have made both requested changes.
