### 2026-02-20 #9711 @rkurchin
🎉 Congrats @tmweigand, and thanks to editor @zhubonan and reviewers @WilkAndy and @saeedmhz, JOSS couldn't function without your efforts!
